Experimental: turn data.frame objects into sc_data

Adapter function to initialize sc_data objects based on a data.frame. Currently, there are heavy restrictions on how the data should look like to work with this function - see Details. This function is not exported.

df_table(x, label = "data")

Arguments

x

a data.frame or an object that is compatible with as.data.frame()

label

a character vector of length one that will be used as a display label for the dataset.

Details

Codes will be automatically assigned to each field, measure and level in the data. Allowed column types are numeric, integer, factor and character.

There are several improvements that need to be made before this will be an exported function

sc_tabulate() currently assumes that each combination of field values occurs once in the data. This is not necessarily the case here.

Time columns are not supported.

Examples

STATcubeR:::df_table(iris)
#> <df_table>
#>   Inherits from: <sc_data>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     data: active binding
#>     field: function (i = 1) 
#>     initialize: function (data, meta, fields) 
#>     language: active binding
#>     meta: active binding
#>     recode: active binding
#>     tabulate: function (...) 
#>     total_codes: function (...) 
#>   Private:
#>     lang: NULL
#>     p_data: sc_tibble, tbl, data.frame
#>     p_fields: list
#>     p_meta: list
#>     recoder: sc_recoder, R6
#>     version: 0.2.8