List available Opendata datasets

od_list() returns a data.frame containing all datasets published at data.statistik.gv.at

od_list(unique = TRUE)

Arguments

unique

some datasets are pulbished under multiple groups. They will only be listed once with the first group they appear in unless this parameter is set to FALSE.

Value

a data.frame with two columns

"category": Grouping under which a dataset is listed

"id": Name of the dataset which can later be used in od_table()

"label": Description of the dataset

Examples

df <- od_list()
df
#> # A data frame: 272 × 3
#>    category    id                                         label                 
#>    <chr>       <chr>                                      <chr>                 
#>  1 Arbeit      OGD_konjunkturmonitor_KonMon_1             Konjunkturmonitor     
#>  2 Arbeit      OGD_veste309_Veste309_1                    Verdienststrukturerhe…
#>  3 Arbeit      OGD_veste307_Veste307_1                    Verdienststrukturerhe…
#>  4 Arbeit      OGD_veste305_Veste305_1                    Verdienststrukturerhe…
#>  5 Arbeit      OGD_veste303_Veste203_1                    Verdienststrukturerhe…
#>  6 Arbeit      OGD_aeapp_biber_abschl_ext_BIBER_ABSCHL_1  BibEr Monitoring 2020…
#>  7 Arbeit      OGD_aest_erv_statw_haupt_ERV_STATWECHSEL_1 Erwerbsverläufe - Sta…
#>  8 Bevölkerung OGD_f1824_AKE1824HH_1                      Privathaushalte (Mikr…
#>  9 Bevölkerung OGD_rate_kalwo_GEST_KALWOCHE_STR_100       Altersstandardisierte…
#> 10 Bevölkerung OGD_gest_kalwo_GEST_KALWOCHE_100           Gestorbene in Österre…
#> # … with 262 more rows
subset(df, category == "Bildung und Forschung")
#> # A data frame: 38 × 3
#>    category              id                            label                    
#>  * <chr>                 <chr>                         <chr>                    
#>  1 Bildung und Forschung OGD_unistud5_ext_UNI_STUD5_1  Studierende in Österreich
#>  2 Bildung und Forschung OGD_unistud4_ext_UNI_STUD4_2  Studierende an öffentlic…
#>  3 Bildung und Forschung OGD_unistud3_ext_UNI_STUD3_1  Studierende an öffentlic…
#>  4 Bildung und Forschung OGD_unistud2_ext_UNI_STUD2_1  Studierende an öffentlic…
#>  5 Bildung und Forschung OGD_unistud1_ext_UNI_STUD1_1  Studien an öffentlichen …
#>  6 Bildung und Forschung OGD_uniabs_ext_UNI_OEFF_ABS_1 Studienabschlüsse an öff…
#>  7 Bildung und Forschung OGD_fhsstud_ext_FHS_S_1       Studien an Fachhochschul…
#>  8 Bildung und Forschung OGD_fhsabs_ext_FHS_A_1        Studienabschlüsse an Fac…
#>  9 Bildung und Forschung OGD_uptstud_ext_UPT_S_1       Studierende an Privatuni…
#> 10 Bildung und Forschung OGD_uptabs_ext_UPT_A_1        Studienabschlüsse an Pri…
#> # … with 28 more rows