Get OGD revisions

Use the /revision endpoint of the OGD server to get a list of all datasets that have changed since a certain timestamp.

Usage

od_revisions(since = NULL, exclude_ext = TRUE, server = "ext")

Arguments

since

(optional) A timestamp. If supplied, only datasets updated later will be returned. Otherwise, all datasets are returned. Can be in either one of the following formats

a native R time type that is compatible with strftime() such as the return values of Sys.Date(), Sys.time() and file.mtime().

a string of the form YYYY-MM-DD to specify a day.

a string of the form YYYY-MM-DDThh:mm:ss to specify a day and a time.

exclude_ext

If TRUE (default) exclude all results that have OGDEXT_ as a prefix

server

the open data server to use. Either ext for the external server (the default) or red for the editing server. The editing server is only accessible for employees of Statistics Austria

Value

a character vector with dataset ids

Examples

# get all datasets (including OGDEXT_*)
ids <- od_revisions(exclude_ext = FALSE)
ids
#> 435 datasets are available ([2024-10-10 10:48:18])
#> ids: OGDEXT_AEST_GEMTAB_1, OGDEXT_AMB_1, OGDEXT_BINNENWAND_1, …, OGD_watlas8_WATLAS_8, and OGD_zlf_komm_ZLF_KOM_1
sample(ids, 6)
#> [1] "OGD_desi_unt_ws_DESI_UNT_WS_1" "OGD_vpi20_VPI_2020_1"         
#> [3] "OGD_1531kn2_Aussenhandel_4"    "OGD_tli06wk10_TLI_104"        
#> [5] "OGD_uniabs_ext_UNI_OEFF_ABS_1" "OGDEXT_KLASSDB_PRODCOM2021_1" 

# get all the datasets since the fifteenth of august
od_revisions("2022-09-15")
#> 295 changes between
#>                 [2022-09-15] and
#>                 [2024-10-10 10:48:18]
#> ids: OGD_1531kn2_Aussenhandel_4, OGD_1905fue_FUE_B1905FUE_1, OGD__steuer_est_ab_2008_altgesch_EST_2_2, …, OGD_watlas8_WATLAS_8, and OGD_zlf_komm_ZLF_KOM_1