Create a table-instance from an open-data dataset

od_table(id) returns an R6-class object containing all relevant data and metadata from https://data.statistik.gv.at/data/

od_table(id, language = c("en", "de"))

Arguments

id

the id of the data-set that should be accessed

language

language to be used for labeling. "en" or "de"

Value

The returned objects is of class sc_table and inherits several parsing methods from sc_data. See od_table_class for the full class documentation.

Components

Component	Corresponding File on Server

$data 	https://data.statistik.gv.at/data/${id}.csv

$header 	https://data.statistik.gv.at/data/${id}_HEADER.csv

$field(code) 	https://data.statistik.gv.at/data/${id}_${code}.csv

$json 	https://data.statistik.gv.at/ogd/json?dataset=${id}

Examples

x <- od_table("OGD_krebs_ext_KREBS_1")

## metadata
x
#> An object of class od_table
#> 
#> Dataset     Cancer statistics by reporting year, province of residence and
#>             localisation of cancer 
#> Measures    Number of records F-KRE 
#> Fields      Tumore ICD/10 3-Steller, Reporting year, Province of residence, Sex 
#> 
#> Request     2021-09-22 15:39:18 
#> STATcubeR   0.2.8 
x$meta
#> $source
#> # STATcubeR metadata: 1 x 7
#>   code                  label                                              lang 
#>   <chr>                 <chr>                                              <chr>
#> 1 OGD_krebs_ext_KREBS_1 Cancer statistics by reporting year, province of … en   
#> # … with 4 more columns: 'label_de', 'label_en', 'requested', 'scr_version'
#> 
#> $measures
#> # STATcubeR metadata: 1 x 5
#>   code  label                     NAs
#>   <chr> <chr>                   <int>
#> 1 F-KRE Number of records F-KRE     0
#> # … with 2 more columns: 'label_de', 'label_en'
#> 
#> $fields
#> # STATcubeR metadata: 4 x 7
#>   code               label                   total_code
#>   <chr>              <chr>                   <chr>     
#> 1 C-TUM_ICD10_3ST-0  Tumore ICD/10 3-Steller NA        
#> 2 C-BERJ-0           Reporting year          NA        
#> 3 C-BUNDESLAND-0     Province of residence   NA        
#> 4 C-KRE_GESCHLECHT-0 Sex                     NA        
#> # … with 4 more columns: 'label_de', 'label_en', 'nitems', 'type'
#> 
x$field("Sex")
#> # STATcubeR metadata: 2 x 7
#>   code         label  parsed
#>   <chr>        <chr>  <chr> 
#> 1 GESCHLECHT-1 male   male  
#> 2 GESCHLECHT-2 female female
#> # … with 4 more columns: 'label_de', 'label_en', 'visible', 'order'
x$field(3)
#> # STATcubeR metadata: 9 x 7
#>   code         label           parsed         
#>   <chr>        <chr>           <chr>          
#> 1 BUNDESLAND-1 "Burgenland "   "Burgenland "  
#> 2 BUNDESLAND-2 "Carinthia"     "Carinthia"    
#> 3 BUNDESLAND-3 "Lower Austria" "Lower Austria"
#> 4 BUNDESLAND-4 "Upper Austria" "Upper Austria"
#> 5 BUNDESLAND-5 "Salzburg"      "Salzburg"     
#> 6 BUNDESLAND-6 "Styria"        "Styria"       
#> 7 BUNDESLAND-7 "Tyrol"         "Tyrol"        
#> 8 BUNDESLAND-8 "Vorarlberg"    "Vorarlberg"   
#> 9 BUNDESLAND-9 "Vienna"        "Vienna"       
#> # … with 4 more columns: 'label_de', 'label_en', 'visible', 'order'

## data
x$data
#> # A STATcubeR tibble: 45,059 x 5
#>    `C-TUM_ICD10_3ST-0` `C-BERJ-0` `C-BUNDESLAND-0` `C-KRE_GESCHLECHT-0` `F-KRE`
#>  * <fct>               <fct>      <fct>            <fct>                  <int>
#>  1 TUM_ICD10_3ST-C00   BERJ-1983  BUNDESLAND-1     GESCHLECHT-1               2
#>  2 TUM_ICD10_3ST-C00   BERJ-1983  BUNDESLAND-2     GESCHLECHT-1               8
#>  3 TUM_ICD10_3ST-C00   BERJ-1983  BUNDESLAND-2     GESCHLECHT-2               2
#>  4 TUM_ICD10_3ST-C00   BERJ-1983  BUNDESLAND-3     GESCHLECHT-1               6
#>  5 TUM_ICD10_3ST-C00   BERJ-1983  BUNDESLAND-3     GESCHLECHT-2               2
#>  6 TUM_ICD10_3ST-C00   BERJ-1983  BUNDESLAND-4     GESCHLECHT-1              12
#>  7 TUM_ICD10_3ST-C00   BERJ-1983  BUNDESLAND-4     GESCHLECHT-2               2
#>  8 TUM_ICD10_3ST-C00   BERJ-1983  BUNDESLAND-5     GESCHLECHT-1               4
#>  9 TUM_ICD10_3ST-C00   BERJ-1983  BUNDESLAND-6     GESCHLECHT-1               5
#> 10 TUM_ICD10_3ST-C00   BERJ-1983  BUNDESLAND-6     GESCHLECHT-2               2
#> # … with 45,049 more rows
x$tabulate()
#> # A STATcubeR tibble: 45,059 x 5
#>    `Tumore ICD/10 3-S… `Reporting year` `Province of res… Sex   `Number of reco…
#>  * <fct>               <date>           <fct>             <fct>            <int>
#>  1 <C00> Bösartige Ne… 1983-01-01       "Burgenland "     male                 2
#>  2 <C00> Bösartige Ne… 1983-01-01       "Carinthia"       male                 8
#>  3 <C00> Bösartige Ne… 1983-01-01       "Carinthia"       fema…                2
#>  4 <C00> Bösartige Ne… 1983-01-01       "Lower Austria"   male                 6
#>  5 <C00> Bösartige Ne… 1983-01-01       "Lower Austria"   fema…                2
#>  6 <C00> Bösartige Ne… 1983-01-01       "Upper Austria"   male                12
#>  7 <C00> Bösartige Ne… 1983-01-01       "Upper Austria"   fema…                2
#>  8 <C00> Bösartige Ne… 1983-01-01       "Salzburg"        male                 4
#>  9 <C00> Bösartige Ne… 1983-01-01       "Styria"          male                 5
#> 10 <C00> Bösartige Ne… 1983-01-01       "Styria"          fema…                2
#> # … with 45,049 more rows

## tabulation: see `?sc_tabulate` for more examples
x$tabulate("Reporting year", "Sex")
#> # A STATcubeR tibble: 72 x 3
#>    `Reporting year` Sex    `Number of records F-KRE`
#>  * <date>           <fct>                      <int>
#>  1 1983-01-01       male                       14492
#>  2 1983-01-01       female                     17478
#>  3 1984-01-01       male                       14798
#>  4 1984-01-01       female                     17456
#>  5 1985-01-01       male                       14555
#>  6 1985-01-01       female                     17448
#>  7 1986-01-01       male                       14626
#>  8 1986-01-01       female                     17241
#>  9 1987-01-01       male                       14862
#> 10 1987-01-01       female                     17845
#> # … with 62 more rows

## switch language
x$language <- "de"
x
#> An object of class od_table
#> 
#> Dataset     Krebsstatistik 
#> Measures    Anzahl der Datensätze F-KRE 
#> Fields      Tumore ICD/10 3-Steller, Berichtsjahr, Bundesland, Geschlecht 
#> 
#> Request     2021-09-22 15:39:18 
#> STATcubeR   0.2.8 
x$tabulate()
#> # A STATcubeR tibble: 45,059 x 5
#>    `Tumore ICD/10 3-Ste… Berichtsjahr Bundesland  Geschlecht `Anzahl der Datens…
#>  * <fct>                 <date>       <fct>       <fct>                    <int>
#>  1 <C00> Bösartige Neub… 1983-01-01   Burgenland  männlich                     2
#>  2 <C00> Bösartige Neub… 1983-01-01   Kärnten     männlich                     8
#>  3 <C00> Bösartige Neub… 1983-01-01   Kärnten     weiblich                     2
#>  4 <C00> Bösartige Neub… 1983-01-01   Niederöste… männlich                     6
#>  5 <C00> Bösartige Neub… 1983-01-01   Niederöste… weiblich                     2
#>  6 <C00> Bösartige Neub… 1983-01-01   Oberösterr… männlich                    12
#>  7 <C00> Bösartige Neub… 1983-01-01   Oberösterr… weiblich                     2
#>  8 <C00> Bösartige Neub… 1983-01-01   Salzburg    männlich                     4
#>  9 <C00> Bösartige Neub… 1983-01-01   Steiermark  männlich                     5
#> 10 <C00> Bösartige Neub… 1983-01-01   Steiermark  weiblich                     2
#> # … with 45,049 more rows

## other interesting tables
od_table("OGD_veste309_Veste309_1")
#> An object of class od_table
#> 
#> Dataset     Structure of Earnings Survey (SES) 2018 Gross hourly earnings in
#>             EUR by citizenship, region (NUTS 2) and form of
#>             employment 
#> Measures    Arithmetic mean, 1st quartile, 2nd quartile (median), 3rd quartile,
#>             Number of employees 
#> Fields      Sex, Citizenship, Region (NUTS2), Form of employment 
#> 
#> Request     2021-09-22 15:39:18 
#> STATcubeR   0.2.8 
od_table("OGD_konjunkturmonitor_KonMon_1")
#> An object of class od_table
#> 
#> Dataset     Economic Trend Monitor 
#> Measures    Production Index Industry (wd; 2015=100), Technical total
#>             production Industry (in 1.000 €), Turnover Index
#>             Industry (2015=100), Turnover Industry (in 1.000 €),
#>             Index of new orders Industry (2015=100), Index of
#>             persons employed Industry (2015=100), Persons employed
#>             Industry, Productivity Index Industry per employee
#>             (2015=100), Productivity Index Industry per hours
#>             worked (2015=100), Industrial Output Price Index
#>             (2015=100), ... 
#> Fields      reporting period, value indication 
#> 
#> Request     2021-09-22 15:39:18 
#> STATcubeR   0.2.8 
od_table("OGD_krankenbewegungen_ex_LEISTUNGEN_1")
#> An object of class od_table
#> 
#> Dataset     Medical procedures during inpatient stays since 1989 by patient
#>             characteristics (number of medical procedures) 
#> Measures    Medical procedures 
#> Fields      Year of discharge, Sex, Age (four classes), NUTS-2 region (place of
#>             residence), Medical procedures - subchapters 
#> 
#> Request     2021-09-22 15:39:18 
#> STATcubeR   0.2.8 
od_table("OGD_f1741_HH_Proj_1")
#> An object of class od_table
#> 
#> Dataset     Household forecast 
#> Measures    Private households at the end of the year, Annual average of
#>             private households 
#> Fields      Time, Province (NUTS 2-Einheit) <9>, Type of household <2>, Age of
#>             household representative in 10-year groups <7>, Sex of
#>             household reference person <2> 
#> 
#> Request     2021-09-22 15:39:18 
#> STATcubeR   0.2.8 
od_table("OGD_veste303_Veste203_1")
#> An object of class od_table
#> 
#> Dataset     Structure of Earnings Survey (SES) 2014 Gross hourly earnings in
#>             EUR by characteristics of the enterprise 
#> Measures    Arithmetic mean, 1st quartile, 2nd quartile (median), 3rd quartile,
#>             Number of employees 
#> Fields      ÖNACE 2008 (NACE Rev.2), Sex, Regions (Nuts1), Size of the
#>             enterprise 
#> 
#> Request     2021-09-22 15:39:18 
#> STATcubeR   0.2.8