Create a table-instance from an open-data dataset

R6 Class open data datasets.

Super class

STATcubeR::sc_data -> od_table

Active bindings

json

parsed version of https://data.statistik.gv.at/ogd/json?dataset=${id}

header

parsed version of https://data.statistik.gv.at/data/${id}_HEADER.csv.

Similar contents can be found in $meta.

resources

lists all files downloaded from the server to construct this table

od_server

The server used for initialization (see to ?od_table)

Methods

Public methods

od_table_class$new()

od_table_class$browse()

Inherited methods

STATcubeR::sc_data$field()

STATcubeR::sc_data$tabulate()

STATcubeR::sc_data$total_codes()

Method new()

This class is not exported. Use od_table() to initialize objects of class od_table.

Usage

od_table_class$new(id, language = NULL, server = "ext")

Arguments

id

the id of the dataset that should be accessed

language

language to be used for labeling. "en" or "de"

server

the OGD-Server server to be used

Method browse()

open the metadata for the dataset in a browser

Usage

od_table_class$browse()