Turn od_table objects into tidy data frames

od_tabulate() extracts the data in the table and turns it into a tidy data.frame. It is basically a copy of sc_tabulate().

the ... argument decides which measures and/or fields should be included in the output. If no measures are given, all measures are included. The same is true for fields.

od_tabulate(table, ..., .list = NULL, raw = FALSE, parse_time = TRUE)

Arguments

table
An object of class od_table

...
Names of measures and/or fields

.list
allows to define the arguments for ... as a character vector

raw
If FALSE (the default), apply labeling to the dataset. Otherwise, return codes.

parse_time
Should time variables be converted into a Date format? Ignored if raw is set to TRUE.

Details

Aggregation is done as follows

First, all columns that priovide a total code via table$total_codes() will be used to filter for column == total_code or column != toal_code

Then, the remaining data is aggregated using rowsum()

The ellipsis (...) supports partial matching of codes, english labels and german labels. See Examples

Examples

table <- od_table("OGD_veste309_Veste309_1")

#> Error: HEADER and attribute description do not match

# no arguments -> same output as `table$data`
od_tabulate(table)

#> Error in od_tabulate(table): inherits(table, "od_table") is not TRUE

# provide some fields -> aggregate to keep only these fields
od_tabulate(table, "Sex", "Citizenship")

#> Error in od_tabulate(table, "Sex", "Citizenship"): inherits(table, "od_table") is not TRUE

# provide some measures -> drop all other measures from the output
od_tabulate(table, "Arithmetic mean")

#> Error in od_tabulate(table, "Arithmetic mean"): inherits(table, "od_table") is not TRUE

# mixture of measures and fields  -> keep exactly those columns
od_tabulate(table, "Sex", "Arithmetic mean")

#> Error in od_tabulate(table, "Sex", "Arithmetic mean"): inherits(table, "od_table") is not TRUE

## define total codes
table$total_codes(
  `C-A11-0` = "A11-1",
  `C-STAATS-0` = "STAATS-9",
  `C-VEBDL-0` = "VEBDL-10",
  `C-BESCHV-0` = "BESCHV-1"
)

#> Error in table$total_codes: object of type 'closure' is not subsettable

## alternatively, use partial matching to define totals
table$total_codes(
  Sex = "Sum total",
  Citizenship = "Total",
  Region = "Total",
  `Form of employment` = "Total"
)

#> Error in table$total_codes: object of type 'closure' is not subsettable

# filter for totals in `Region (NUTS2)` and `Form of employment`. Drop totals
# in `Sex` and `Citizenship`.
od_tabulate(table, "Sex", "Citizenship")

#> Error in od_tabulate(table, "Sex", "Citizenship"): inherits(table, "od_table") is not TRUE

## switch language
table$language <- "de"

#> Error in table$language <- "de": object of type 'closure' is not subsettable

od_tabulate(table, "Sex", "Citizenship")

#> Error in od_tabulate(table, "Sex", "Citizenship"): inherits(table, "od_table") is not TRUE

## `...` matches for codes, english labels and german labels
od_tabulate(table, "C-A11-0", "Citizenship", "2. Quartil (Median)")

#> Error in od_tabulate(table, "C-A11-0", "Citizenship", "2. Quartil (Median)"): inherits(table, "od_table") is not TRUE

## Keep totals in the output by removing total codes
od_tabulate(table, "C-A11-0")      # -> 2 rows: "male" "female"

#> Error in od_tabulate(table, "C-A11-0"): inherits(table, "od_table") is not TRUE

table$total_codes(`C-A11-0` = NA)

#> Error in table$total_codes: object of type 'closure' is not subsettable

od_tabulate(table, "C-A11-0")      # -> 3 rows: "total", "male", "female"

#> Error in od_tabulate(table, "C-A11-0"): inherits(table, "od_table") is not TRUE

## table$tabulate(...) is an alias for sc_tabulate(table, ...)
table$tabulate("C-A11-0")

#> Error in table$tabulate: object of type 'closure' is not subsettable