Cache responses from the STATcube REST API

Functions to cache requested resources in the directory ~/.STATcubeR_cache and reuse them in calls to sc_table(), sc_table_custom() sc_schema() and so forth. These functions are designed for testing and documentation and should not be regarded as part of the STATcubeR interface. The caching logic is likely to change in the future in which case sc_cache_clear() is required to purge old cache entries.

Usage

## enable caching for the current R session
sc_cache_enable(verbose)

## disable caching for the current R session
sc_cache_disable()

sc_cache_enabled()

sc_cache_dir(dir = NULL)

sc_cache_files(x)

## remove all files from the cache
sc_cache_clear()

Arguments

verbose

print instuctions on how to set up caching persistently via environment variables?

dir

a chace directory

x

an object of class sc_table or sc_schema

Details

Caching can be set up using environment variables. To set up a persistent cache for both Open Data and the REST API, the following lines in .Renviron can be used.

STATCUBE_KEY       = YOUR_API_KEY_GOES_HERE
STATCUBE_CACHE     = TRUE
OD_CACHE_DIR       = "~/.cache/STATcubeR/open_data/"
STATCUBE_CACHE_DIR = "~/.cache/STATcubeR/api/"

Note that the caches are always used and there is no check to verify if the resources are unchanged in the server. Caching is not implemented for the endpoints sc_info() and sc_rate_limit_table().

Functions

sc_cache_enabled(): informs wether the cache is currently enabled

sc_cache_dir(): get/set the directory used for caching

sc_cache_files(): get the cache file associated with an object