Get metadata for a STATcube table

Functions to extract metadata from a sc_table object.

sc_meta(response)

sc_meta_field(response, i = 1)

Arguments

response
An object of class sc_table

i
index of the field, for which further metadata are desired

See also

Examples

my_response <- sc_example("population_timeseries.json") %>%
  sc_table()
sc_meta(my_response)

#> $database
#>                                                label         code
#> 1 Population at the beginning of the year since 1982 debevstandjb
#> 
#> $measures
#>    label     code fun precision
#> 1 Number F-ISIS-1 SUM         0
#> 
#> $fields
#>              label        code nitems        type
#> 1     Time section     C-A10-0     39 Time (year)
#> 2       Bundesland    C-BB00-0     10    Category
#> 3 Country of birth C-GEBLAND-0      2    Category
#> 

sc_meta_field(my_response, 1)

#>    label     code     parsed
#> 1   1982 A10-1982 1982-01-01
#> 2   1983 A10-1983 1983-01-01
#> 3   1984 A10-1984 1984-01-01
#> 4   1985 A10-1985 1985-01-01
#> 5   1986 A10-1986 1986-01-01
#> 6   1987 A10-1987 1987-01-01
#> 7   1988 A10-1988 1988-01-01
#> 8   1989 A10-1989 1989-01-01
#> 9   1990 A10-1990 1990-01-01
#> 10  1991 A10-1991 1991-01-01
#> 11  1992 A10-1992 1992-01-01
#> 12  1993 A10-1993 1993-01-01
#> 13  1994 A10-1994 1994-01-01
#> 14  1995 A10-1995 1995-01-01
#> 15  1996 A10-1996 1996-01-01
#> 16  1997 A10-1997 1997-01-01
#> 17  1998 A10-1998 1998-01-01
#> 18  1999 A10-1999 1999-01-01
#> 19  2000 A10-2000 2000-01-01
#> 20  2001 A10-2001 2001-01-01
#> 21  2002 A10-2002 2002-01-01
#> 22  2003 A10-2003 2003-01-01
#> 23  2004 A10-2004 2004-01-01
#> 24  2005 A10-2005 2005-01-01
#> 25  2006 A10-2006 2006-01-01
#> 26  2007 A10-2007 2007-01-01
#> 27  2008 A10-2008 2008-01-01
#> 28  2009 A10-2009 2009-01-01
#> 29  2010 A10-2010 2010-01-01
#> 30  2011 A10-2011 2011-01-01
#> 31  2012 A10-2012 2012-01-01
#> 32  2013 A10-2013 2013-01-01
#> 33  2014 A10-2014 2014-01-01
#> 34  2015 A10-2015 2015-01-01
#> 35  2016 A10-2016 2016-01-01
#> 36  2017 A10-2017 2017-01-01
#> 37  2018 A10-2018 2018-01-01
#> 38  2019 A10-2019 2019-01-01
#> 39  2020 A10-2020 2020-01-01