Create a request against the /schema endpoint

Invoke the /schema endpoint of the STATcube REST API. In case of sc_schema_catalogue(), recurse into all datasets and tables and return a nested list with ids and labels for all resources. For sc_schema_db(), recurse into all valuesets and return a list of all resources available tor the specific dataset. The return values can be displayed as a tree object.

sc_schema(
  resource_id = NULL,
  depth = NULL,
  language = c("en", "de"),
  key = sc_key()
)

# S3 method for sc_schema
print(x, tree = NULL, ..., limit = 30)

sc_schema_catalogue(depth = "folder", language = c("en", "de"), key = sc_key())

sc_schema_db(
  db_id,
  depth = "valueset",
  language = c("en", "de"),
  key = sc_key()
)

Arguments

resource_id
A resource identifier in uid format

depth
If provided, the request will recurse into the given level. For datasets, available options are NULL (no recursion), "folder", "field" and "valueset". For the catalogue, only NULL and "folder" are applicable.

language
The language to be used for labeling. "en" or "de"

key
(string) An API key. To display your key, call sc_browse_preferences().

x
object to be printed

tree
wether to use the data.tree package for printing.

limit, ...
passed to data.tree::print.Node()

db_id
a database id

Printing with data.tree

limit and ... will simply be ignored if tree is set to FALSE, which is the default. The printing via data.tree can take longer than the default implementation because x will need to be converted into a data.tree node. To use data.tree printing permanently, use

options(STATcubeR.print_tree = TRUE)

Examples

my_catalogue <- sc_schema_catalogue()

## print
my_catalogue

#> FOLDER: Extern V9
#> # A data frame: 8 × 2
#>   child                     type  
#>   <chr>                     <chr> 
#> 1 Examples                  FOLDER
#> 2 Publications and services FOLDER
#> 3 Default Tables            FOLDER
#> 4 Statistics                FOLDER
#> 5 ts_tourism                TABLE 
#> 6 krankenbewegungen         TABLE 
#> 7 tourism_ts                TABLE 
#> 8 entlassungen              TABLE 

## access the parsed catalogue
my_catalogue$Statistics$`Labour Market`

#> FOLDER: Labour Market
#> # A data frame: 14 × 2
#>    child                                                      type    
#>    <chr>                                                      <chr>   
#>  1 Job Seeker                                                 FOLDER  
#>  2 Labour Market Policy                                       FOLDER  
#>  3 Working Time                                               FOLDER  
#>  4 Activity Status                                            FOLDER  
#>  5 Employment                                                 FOLDER  
#>  6 Job Vacancies                                              FOLDER  
#>  7 Working hours (Labour Force Survey)                        DATABASE
#>  8 Austrian Micro census - Labour Force Survey Yearly Data    DATABASE
#>  9 Austrian Micro census - Labour Force Survey Quarterly Data DATABASE
#> 10 Microcensus Time Series since 1974                         DATABASE
#> 11 Standardtabelle / Default table (defaulttable_deake001j)   TABLE   
#> 12 Standardtabelle / Default table (defaulttable_deake003j)   TABLE   
#> 13 Standardtabelle / Default table (defaulttable_deake005)    TABLE   
#> 14 Standardtabelle / Default table (defaulttable_deake002q)   TABLE   

my_catalogue$Statistics$`Labour Market`$`Working hours (Labour Force Survey)`

#> DATABASE: Working hours (Labour Force Survey)
#> # Get more info with `sc_schema_db('str:database:deake005')` 

db_schema <- sc_schema_db("deake005")

# printing
db_schema

#> DATABASE: Working hours (Labour Force Survey)
#> # A data frame: 9 × 2
#>   child                                type 
#>   <chr>                                <chr>
#> 1 Factors                              GROUP
#> 2 Facts                                GROUP
#> 3 Datensätze/Records                   GROUP
#> 4 Time (mandatory field)               GROUP
#> 5 Demographic Characteristics          GROUP
#> 6 Employment Characteristics           GROUP
#> 7 Working Time                         GROUP
#> 8 Characteristics of second employment GROUP
#> 9 Geographic Characteristics           GROUP

# access child nodes
db_schema$`Demographic Characteristics`

#> GROUP: Demographic Characteristics
#> # A data frame: 8 × 2
#>   child                                                         type 
#>   <chr>                                                         <chr>
#> 1 Gender                                                        FIELD
#> 2 Age in years-groups                                           FIELD
#> 3 Highest level of training completed (national classification) FIELD
#> 4 Highest level of training completed (ISCED 97)                FIELD
#> 5 Highest level of training completed (ISCED 2011)              FIELD
#> 6 Nationality                                                   FIELD
#> 7 Country of birth                                              FIELD
#> 8 Migration background                                          FIELD

db_schema$`Demographic Characteristics`$Gender$Gender

#> VALUESET: Gender
#> # A data frame: 3 × 2
#>   child                type 
#>   <chr>                <chr>
#> 1 male                 VALUE
#> 2 female               VALUE
#> 3 Not classifiable <0> VALUE

db_schema$`Demographic Characteristics`$Gender$Gender$male

#> VALUE: male

# access the raw response from httr::GET()
my_response <- attr(db_schema, "response")
my_response$headers$date

#> [1] "Wed, 11 Aug 2021 13:59:05 GMT"

my_content <- httr::content(my_response)
my_content$label

#> [1] "Working hours (Labour Force Survey)"

# print with data.tree
print(my_catalogue$Examples, tree = TRUE)

#>                                                        levelName     type
#> 1  Examples                                                        FOLDER
#> 2   ¦--Communes (Demo)                                           DATABASE
#> 3   ¦--People                                                    DATABASE
#> 4   ¦--People multilingual                                       DATABASE
#> 5   ¦--Retail Banking ML en                                      DATABASE
#> 6   ¦--Monitoring                                                   TABLE
#> 7   ¦--Standardtabelle / Default table (defaulttable_depeopleml)    TABLE
#> 8   ¦--Standardtabelle / Default table (defaulttable_dedemo)        TABLE
#> 9   ¦--E-A_nach_Bundeslaendern_dedemo                               TABLE
#> 10  °--Jahre_nach_NUTS_dedemo                                       TABLE