Create a request against the /table endpoint

Send requests against the /table endpoint of the STATcube REST API. The requests can use a json file or a table uri to specify the table. In both cases, an object of class "sc_table" is returned.

sc_table(json_file, language = c("en", "de"), key = sc_key())

sc_example(filename)

# S3 method for sc_table
as.data.frame(
  x,
  ...,
  drop_aggregates = TRUE,
  recode_na = TRUE,
  label_vars = TRUE,
  parse_fields = TRUE
)

sc_table_saved_list()

sc_table_saved(table_uri, language = c("en", "de"), key = sc_key())

Arguments

json_file
path to a json file, which was downloaded via the STATcube gui ("Open Data API Abfrage")

language
The language to be used for labeling. "en" or "de"

key
(string) An API key. To display your key, call sc_browse_preferences().

filename
The name of an example json file.

x
an R object of class sc_table

...
unused

drop_aggregates
remove rows containing aggregates from the table. The default (TRUE) makes the table tidy

recode_na
Recode cells to NA if the corresponding annotations have the value "X" (cross tabulation not allowed).

label_vars
Use labels for column names. Alternatively, codes are used

parse_fields
Parse field columns as character or POSIXct depending on labels? Alternatively, codes are used

table_uri
Identifier of a saved table as returned by sc_table_saved_list()

Value

An object of class sc_table which contains the return value of the httr::POST() request in obj$response. The object also provides member functions to parse this response object.

See also

Examples

if (FALSE) {
my_table <- sc_table(json_file = sc_example("LGR01.json"))

# print
my_table

# get matadata for the table
my_table$meta

# get a data.frame
as.data.frame(my_table)

# get metadata for field 1
my_table$field(1)
}
if (FALSE) {

# get the ids and labels of all saved tables
(saved_tables <- sc_table_saved_list())
table_uri <- saved_tables$id[1]

# get a table based on one of these ids
my_response <- sc_table_saved(table_uri)
as.data.frame(my_response) %>% head()
}