data interface for STATgraph

Convert R objects into a special format for the web application. Currently, the following objects are convertible

od_table: OGD Datasets

sc_table: Responses from the STATcube REST API

data.frame: Requires that all columns are of type factor, character, integer numeric, or Date.

sg_data(x, ...)

Arguments

x
An object to be converted

...
Specifiers for the conversion. Currently ignored.

Value

An object of class sg_data with the following methods

$vars contains codes and labels for all

$times

$measures

$categories

$tabulate(..., .list) aggregates the data. ... should contain the codes from $vars.

Examples

library(STATcubeR)

cancer <- od_table("OGD_krebs_ext_KREBS_1", language = "de")
sg_data(cancer)

pop <- sc_table(json_file = sc_example("population_timeseries.json"))
sg_data(pop)

sg_data(iris)

## basic usage
y <- sg_data(cancer)
y$vars
y$tabulate()

## recode varnames
y$recode_varname("C-BERJ-0", "Jahr")
y$recode_varname("C-TUM_ICD10_3ST-0", "Tumore")
y$recode_varname("F-KRE", "Anzahl")
y$tabulate()

## recode levels
y$recode_levels("C-KRE_GESCHLECHT-0", c("M", "W"))
y$recode_levels("C-BUNDESLAND-0", LETTERS[1:9])
y$tabulate()