Utility Functions

sc_write_json() saves the api request for a table as a json file. The resulting json file can be passed to sc_table()

sc_annotation_legend() lists all annotations occurring in a table together with descriptions of the annotations.

sc_write_json(response, file)

sc_annotation_legend(response)

Arguments

response
an object of class sc_table

file
A filename, under which the json should be saved.

Details

Supporting functions for STATcubeR