Cache management for Open Data

Functions to inspect the contents of the current cache.

od_cache_summary()

od_downloads()

Details

od_cache_summary() provides an overview of all contents of the cache through a data.frame. It hasone row for each dataset and the following columns. All file sizes are given in bytes

id the dataset id

updated the last modified time for ${id}.json

json the file size of ${id}.json

data the file size of ${id}.csv

header the file size of ${id}_HEADER.csv

fields the total file size of all files belonging to fields ({id}_C*.csv).

n_fields the number of field files

od_downloads() shows a download history for the current cache

time a timestamp for the download

file the filename

downloaded the download time in milliseconds

Examples

## make sure the cache is not empty
od_table("OGD_krebs_ext_KREBS_1")
#> An object of class od_table
#> 
#> Dataset     Cancer statistics by reporting year, province of residence and
#>             localisation of cancer 
#> Measures    Number of records F-KRE 
#> Fields      Tumore ICD/10 3-Steller, Reporting year, Province of residence, Sex 
#> 
#> Request     2021-09-17 15:36:37 
#> STATcubeR   0.2.8 
od_table("OGD_veste309_Veste309_1")
#> An object of class od_table
#> 
#> Dataset     Structure of Earnings Survey (SES) 2018 Gross hourly earnings in
#>             EUR by citizenship, region (NUTS 2) and form of
#>             employment 
#> Measures    Arithmetic mean, 1st quartile, 2nd quartile (median), 3rd quartile,
#>             Number of employees 
#> Fields      Sex, Citizenship, Region (NUTS2), Form of employment 
#> 
#> Request     2021-09-17 15:36:37 
#> STATcubeR   0.2.8 

## inspect
od_cache_summary()
#> # A data frame: 37 × 7
#>    id                    updated              json   data header fields n_fields
#>    <chr>                 <dttm>              <dbl>  <dbl>  <dbl>  <dbl>    <int>
#>  1 OGD_aeapp_biber_absc… 2021-08-26 16:17:19  5063 3.57e4    429   1327        4
#>  2 OGD_aest_erv_statw_h… 2021-08-27 13:47:59  6237 8.83e4    476   1700        5
#>  3 OGD_ake002q_AKEQ_1    2021-09-07 12:57:32  2575 1.12e3    115   3585        1
#>  4 OGD_bevstandjbab2002… 2021-08-26 16:29:35  3929 1.78e7    303 102908        4
#>  5 OGD_bevstandjbab2002… 2021-08-13 19:04:33  3929 1.79e7    303 102908        4
#>  6 OGD_bevstandjbab2002… 2021-08-13 19:04:45  3929 1.80e7    303 102908        4
#>  7 OGD_bevstandjbab2002… 2021-09-17 15:02:27  3930 1.82e7    303 102908        4
#>  8 OGD_bevwan020_AUSSEN… 2021-09-08 11:16:25  3756 1.02e7    441  15324        4
#>  9 OGD_bevwan020_AUSSEN… 2021-08-19 17:48:26  3876 2.95e7    544  99475        4
#> 10 OGD_bevwan020_AUSSEN… 2021-08-13 18:59:27  3903 3.73e6    483  10733        4
#> # … with 27 more rows
od_downloads()
#> # A data frame: 786 × 3
#>    time                file                                       downloaded
#>  * <dttm>              <chr>                                           <dbl>
#>  1 2021-09-17 15:02:31 OGD_vgrrgr104_RGR104_1.json                      4.32
#>  2 2021-09-17 15:02:30 OGD_f1741_HH_Proj_2.json                         5.38
#>  3 2021-09-17 15:02:27 OGD_bevstandjbab2002_BevStand_2020.json          6.69
#>  4 2021-09-17 15:02:07 OGD_veste303_Veste203_1.json                     6.32
#>  5 2021-09-17 15:02:07 OGD_f1741_HH_Proj_1.json                         7.45
#>  6 2021-09-17 15:02:07 OGD_krankenbewegungen_ex_LEISTUNGEN_1.json       7.40
#>  7 2021-09-17 15:02:06 OGD_konjunkturmonitor_KonMon_1.json              7.59
#>  8 2021-09-17 15:02:04 OGD_krebs_ext_KREBS_1.json                      47.2 
#>  9 2021-09-17 14:53:50 OGD_veste309_Veste309_1.json                    50.5 
#> 10 2021-09-16 17:16:50 OGD_touextsai_Tour_UA_1.json                     9.18
#> # … with 776 more rows