Cache management for Open Data

Functions to inspect the contents of the current cache.

Usage

od_cache_summary(server = "ext")

od_downloads(server = "ext")

Arguments

server

the OGD-Server to use. "ext" for the external server (the default) or "red" for the editing server

Details

od_cache_summary() provides an overview of all contents of the cache through a data.frame. It hasone row for each dataset and the following columns. All file sizes are given in bytes

id the dataset id

updated the last modified time for ${id}.json

json the file size of ${id}.json

data the file size of ${id}.csv

header the file size of ${id}_HEADER.csv

fields the total file size of all files belonging to fields ({id}_C*.csv).

n_fields the number of field files

od_downloads() shows a download history for the current cache

time a timestamp for the download

file the filename

downloaded the download time in milliseconds

Examples

## make sure the cache is not empty
od_table("OGD_krebs_ext_KREBS_1")
#> An object of class od_table
#> 
#> Dataset     Cancer statistics by reporting year, province of
#>             residence and localisation of cancer 
#> Measures    Number of records F-KRE 
#> Fields      Tumore ICD/10 3-Steller, Reporting year, Province of
#>             residence, Sex 
#> 
#> Request     2022-08-31 18:55:04 
#> STATcubeR   0.4.1 
od_table("OGD_veste309_Veste309_1")
#> An object of class od_table
#> 
#> Dataset     Structure of Earnings Survey (SES) 2018 Gross hourly
#>             earnings in EUR by citizenship, region (NUTS
#>             2) and form of employment 
#> Measures    Arithmetic mean, 1st quartile, 2nd quartile (median), 3rd
#>             quartile, Number of employees 
#> Fields      Sex, Citizenship, Region (NUTS2), Form of employment 
#> 
#> Request     2022-08-31 18:55:04 
#> STATcubeR   0.4.1 

## inspect
od_cache_summary()
#> # A data frame: 57 × 7
#>    id           updated              json   data header fields n_fie…¹
#>    <chr>        <dttm>              <dbl>  <dbl>  <dbl>  <dbl>   <int>
#>  1 OGD_1531kn2… 2022-01-26 08:28:58  3089 1.06e5    277   9435       2
#>  2 OGD_aeapp_b… 2022-07-26 17:35:31  5059 3.96e4    407   1612       4
#>  3 OGD_aest_er… 2021-10-22 17:31:15  6237 8.83e4    476   1700       5
#>  4 OGD_ake002q… 2021-09-07 12:57:32  2575 1.12e3    115   3585       1
#>  5 OGD_bevjahr… 2021-08-10 19:08:18  4412 1.58e6   1444   5262       4
#>  6 OGD_bevstan… 2021-08-26 16:29:35  3929 1.78e7    303 102908       4
#>  7 OGD_bevstan… 2021-08-13 19:04:33  3929 1.79e7    303 102908       4
#>  8 OGD_bevstan… 2021-08-13 19:04:45  3929 1.80e7    303 102908       4
#>  9 OGD_bevstan… 2022-08-31 13:37:15  3897 1.82e7    303 102827       4
#> 10 OGD_bevstan… 2021-11-22 16:04:09  3930 1.82e7    303 102908       4
#> # … with 47 more rows, and abbreviated variable name ¹​n_fields
od_downloads()
#> # A data frame: 1,435 × 3
#>    time                file                                    downl…¹
#>  * <dttm>              <chr>                                     <dbl>
#>  1 2022-08-31 18:55:04 OGD_veste309_Veste309_1.json              11.2 
#>  2 2022-08-31 18:55:04 OGD_krebs_ext_KREBS_1.json                57.8 
#>  3 2022-08-31 13:37:19 OGD_vgrrgr104_RGR104_1.json                5.90
#>  4 2022-08-31 13:37:19 OGD_f1741_HH_Proj_2.json                   4.63
#>  5 2022-08-31 13:37:15 OGD_bevstandjbab2002_BevStand_2020.json    8.18
#>  6 2022-08-31 13:36:39 OGD_veste303_Veste203_1.json               9.24
#>  7 2022-08-31 13:36:38 OGD_f1741_HH_Proj_1.json                   5.14
#>  8 2022-08-31 13:36:38 OGD_krankenbewegungen_ex_LEISTUNGEN_1.…    5.56
#>  9 2022-08-31 13:36:38 OGD_konjunkturmonitor_KonMon_1.json        9.66
#> 10 2022-08-31 13:36:37 OGD_fhsstud_ext_FHS_S_1.json               7.08
#> # … with 1,425 more rows, and abbreviated variable name ¹​downloaded