Cache management for Open Data

Functions to inspect the contents of the current cache.

od_cache_summary()

od_downloads()

Details

od_cache_summary() provides an overview of all contents of the cache through a data.frame. It hasone row for each dataset and the following columns. All file sizes are given in bytes

id the dataset id

updated the last modified time for ${id}.json

json the file size of ${id}.json

data the file size of ${id}.csv

header the file size of ${id}_HEADER.csv

fields the total file size of all files belonging to fields ({id}_C*.csv).

n_fields the number of field files

od_downloads() shows a download history for the current cache

time a timestamp for the download

file the filename

downloaded the download time in milliseconds

Examples

## make sure the cache is not empty
od_table("OGD_krebs_ext_KREBS_1")
#> An object of class od_table
#> 
#> Dataset     Cancer statistics by reporting year, province of residence and
#>             localisation of cancer 
#> Measures    Number of records F-KRE 
#> Fields      Tumore ICD/10 3-Steller, Reporting year, Province of residence, Sex 
#> 
#> Request     2021-08-18 18:22:20 
#> STATcubeR   0.2.4 (@4d989e1) 
od_table("OGD_veste309_Veste309_1")
#> An object of class od_table
#> 
#> Dataset     Structure of Earnings Survey (SES) 2018 Gross hourly earnings in
#>             EUR by citizenship, region (NUTS 2) and form of
#>             employment 
#> Measures    Arithmetic mean, 1st quartile, 2nd quartile (median), 3rd quartile,
#>             Number of employees 
#> Fields      Sex, Citizenship, Region (NUTS2), Form of employment 
#> 
#> Request     2021-08-18 18:22:20 
#> STATcubeR   0.2.4 (@4d989e1) 

## inspect
od_cache_summary()
#> # A data frame: 271 × 7
#>    id                    updated              json   data header fields n_fields
#>    <chr>                 <dttm>              <dbl>  <dbl>  <dbl>  <dbl>    <int>
#>  1 OGD_1531kn2_Aussenha… 2021-08-13 00:53:09  3089 106427    277   9435        2
#>  2 OGD__steuer_lst_ab_2… 2021-08-13 00:53:09  6528  83665   4259   1695        5
#>  3 OGD__steuer_lst_ab_2… 2021-08-13 00:53:09  6208 305349   4297   2653        4
#>  4 OGD__steuer_lst_ab_2… 2021-08-13 00:53:09  6513  42023   4259   1676        5
#>  5 OGD__steuer_lst_ab_2… 2021-08-13 00:53:09  6193 153282   4297   2634        4
#>  6 OGD__steuer_lue_2015… 2021-08-13 00:53:09  4862 269849   2263   3436        3
#>  7 OGD_aeapp_biber_absc… 2021-08-17 12:52:47  5063  35682    429   1327        4
#>  8 OGD_aest_erv_statw_h… 2021-08-18 00:11:58  6237  88258    476   1700        5
#>  9 OGD_ahlgrhs_Aussenha… 2021-08-13 00:53:09  3318  13886    270   1276        3
#> 10 OGD_ahlgrhs_Aussenha… 2021-08-13 00:53:09  2974  28513    225   7681        2
#> # … with 261 more rows
od_downloads()
#> # A data frame: 4,335 × 3
#>    time                file                                           downloaded
#>  * <dttm>              <fct>                                               <dbl>
#>  1 2021-08-18 17:50:11 OGD_veste303_Veste203_1.json                         21.4
#>  2 2021-08-18 17:50:11 OGD_f1741_HH_Proj_1.json                             11.5
#>  3 2021-08-18 17:50:10 OGD_krankenbewegungen_ex_LEISTUNGEN_1.json           12.5
#>  4 2021-08-18 17:50:10 OGD_konjunkturmonitor_KonMon_1_C-QUOTE-1.csv         10.7
#>  5 2021-08-18 17:50:10 OGD_konjunkturmonitor_KonMon_1_C-KMMONAT-0.csv       11.9
#>  6 2021-08-18 17:50:10 OGD_konjunkturmonitor_KonMon_1_HEADER.csv            12.3
#>  7 2021-08-18 17:50:10 OGD_konjunkturmonitor_KonMon_1.csv                  556. 
#>  8 2021-08-18 17:50:10 OGD_konjunkturmonitor_KonMon_1.json                 284. 
#>  9 2021-08-18 17:50:08 OGD_veste309_Veste309_1.json                        137. 
#> 10 2021-08-18 17:50:08 OGD_krebs_ext_KREBS_1.json                          103. 
#> # … with 4,325 more rows