Get a catalogue for OGD datasets

EXPERIMENTAL This function parses several json metadata files at once and combines them into a data.frame so the datasets can easily be filtered based on categorizations, tags, number of classifications, etc.

Usage

od_catalogue(server = "ext", local = TRUE)

Arguments

server

the OGD-server to be used. "ext" (the default) for the external server or prod for the production server

local

If TRUE (the default), the catalogue is created based on cached json metadata. Otherwise, the cache is updated prior to creating the catalogue using a "bulk-download" for metadata files.

Details

The naming, ordering and choice of the columns is likely to change. Currently, the following columns are provided.

Column	Type	Description

title	chr	Title of the dataset

measures	int	Number of measure variables

fields	int	Number of classification fields

modified	datetime	Timestamp when the dataset was last modified

created	datetime	Timestamp when the dataset was created

database	chr	ID of the corresponding STATcube database

title_en	chr	English title

notes	chr	Description for the dataset

frequency	chr	How often is the dataset updated?

category	chr	Category of the dataset

tags	list<chr>	tags assigned to the dataset

json	list<od_json>	Full json metadata

The type datetime refers to the POSIXct format as returned by Sys.time(). The last column "json" containes the full json metadata as returned by od_json().

Examples

catalogue <- od_catalogue()
catalogue
#> # A data frame: 315 × 13
#>    title  measu…¹ fields modified            created             id   
#>    <chr>    <int>  <int> <dttm>              <dttm>              <chr>
#>  1 Außen…       2      2 2022-06-28 13:38:12 2018-07-02 13:46:30 OGD_…
#>  2 FuE- …       6      2 2022-09-16 14:23:21 2018-01-12 20:00:00 OGD_…
#>  3 Einko…     103      4 2022-07-08 20:04:25 2021-10-15 20:00:00 OGD_…
#>  4 Einko…     103      4 2022-07-08 20:04:01 2021-10-15 20:00:00 OGD_…
#>  5 Einko…     103      4 2022-07-08 20:04:37 2021-10-15 20:00:00 OGD_…
#>  6 Körpe…      19      4 2022-07-08 20:04:12 2021-10-18 20:00:00 OGD_…
#>  7 Körpe…      19      4 2022-07-08 20:04:12 2021-10-18 20:00:00 OGD_…
#>  8 Lohns…      58      4 2022-07-07 20:04:01 2021-10-15 20:00:00 OGD_…
#>  9 Lohns…      58      4 2022-07-07 20:03:44 2021-10-15 20:00:00 OGD_…
#> 10 Lohns…      36      5 2017-12-20 09:06:52 2016-12-19 21:32:56 OGD_…
#> # … with 305 more rows, 7 more variables: database <chr>,
#> #   title_en <chr>, notes <chr>, update_frequency <chr>,
#> #   tags <I<list>>, categorization <chr>, json <I<list>>, and
#> #   abbreviated variable name ¹​measures
catalogue$update_frequency %>% table()
#> .
#>  halbjährlich      jährlich     monatlich   nach Bedarf nicht geplant 
#>             5           149            84            28            16 
#> quartalsweise   wöchentlich 
#>            29             4 
catalogue$categorization %>% table()
#> .
#>                      Arbeit                 Bevölkerung 
#>                           7                          85 
#>       Bildung und Forschung Finanzen und Rechnungswesen 
#>                          33                          26 
#>   Gesellschaft und Soziales                  Gesundheit 
#>                          12                           6 
#>                      Umwelt         Verkehr und Technik 
#>                           1                          16 
#>    Wirtschaft und Tourismus 
#>                         129 
catalogue[catalogue$categorization == "Gesundheit", 1:4]
#> # A data frame: 6 × 4
#>   title                             measu…¹ fields modified           
#> * <chr>                               <int>  <int> <dttm>             
#> 1 Spitalsentlassungen aus Akutkran…       2      5 2021-07-28 15:33:51
#> 2 Spitalsentlassungen aus Akutkran…       2      4 2022-09-08 10:16:39
#> 3 Spitalsentlassungen aus Krankena…       2      5 2021-07-28 15:33:51
#> 4 Spitalsentlassungen aus Krankena…       2      4 2022-09-08 10:16:39
#> 5 Medizinische Leistungen während …       1      5 2022-09-08 10:16:39
#> 6 Krebsstatistik                          1      4 2022-09-06 14:51:39
#> # … with abbreviated variable name ¹​measures
catalogue[catalogue$measures >= 70, 1:3]
#> # A data frame: 16 × 3
#>    title                                                measu…¹ fields
#>  * <chr>                                                  <int>  <int>
#>  1 Einkommensteuerstatistik ab 2008 nach Stufen des zu…     103      4
#>  2 Einkommensteuerstatistik ab 2008 nach Bundesländern…     103      4
#>  3 Einkommensteuerstatistik ab 2008 nach Bundesländern…     103      4
#>  4 Index der Beschäftigten insgesamt ÖNACE 2008 (Basis…      74      1
#>  5 BKI: Baukostenindex für den Brückenbau (2015=100)         79      1
#>  6 BKI: Baukostenindex für den Brückenbau (2020=100)         79      1
#>  7 BKI: Baukostenindex für den Straßenbau (2015=100)         79      1
#>  8 BKI: Baukostenindex für den Straßenbau (2020=100)         79      1
#>  9 Bruttoverdiensteindex (Basis 2015)                        74      1
#> 10 Bruttoverdiensteindex (Basis 2015), EU-harmonisiert…      74      2
#> 11 Arbeitsvolumensindex (Basis 2015), EU-harmonisierte…      75      2
#> 12 Produktionsindex Nace Rev.2 (Basis 2015)                  75      3
#> 13 Umsatzindex Nace Rev.2 (Basis 2015), EU-harmonisier…      75      3
#> 14 Konjunkturmonitor                                         86      2
#> 15 Öffentliche Finanzen ab 1995, ESVG 2010                   76      2
#> 16 Nicht-finanzielle Sektorkonten nach ESVG 2010            124      2
#> # … with abbreviated variable name ¹​measures
catalogue$json[[1]]
#> Außenhandelsdaten Jährlich
#> 
#> Euro Werte der Einfuhren bzw. Ausfuhren nach Ursprungs- bzw.
#> Bestimmungsländern, jährliche Zeitreihe ab 2007
#> 
#> Measures: Einfuhr Wert in Euro, Ausfuhr Wert in Euro
#> Fields: Jahr, Ursprungs- bzw. Bestimmungsland
#> Updated: 2022-06-28 13:38:12
#> Tags: Einfuhren, Ausfuhren, Außenhandel, Ursprungsland,
#>   Bestimmungsland
#> Categories: Wirtschaft und Tourismus
catalogue$database %>% head()
#> [1] "de1531kn2"                      "de1905fue"                     
#> [3] "de_steuer_est_ab_2008_altgesch" "de_steuer_est_ab_2008_bl_eink" 
#> [5] "de_steuer_est_ab_2008_blgesch"  "de_steuer_kst"