List available Opendata datasets

od_list() returns a data.frame containing all datasets published at data.statistik.gv.at

Usage

od_list(unique = TRUE, server = c("ext", "red"))

Arguments

unique

some datasets are published under multiple groups. They will only be listed once with the first group they appear in unless this parameter is set to FALSE.

server

the open data server to use. Either ext for the external server (the default) or red for the editing server. The editing server is only accessible for employees of Statistics Austria

Value

a data.frame with two columns

"category": Grouping under which a dataset is listed

"id": Name of the dataset which can later be used in od_table()

"label": Description of the dataset

Examples

df <- od_list()
df
#> # A tibble: 384 × 3
#>    category      id                                   label           
#>    <chr>         <chr>                                <chr>           
#>  1 Neueste Daten OGD_bevjahresanf_PR_BEVJA_4          Bevölkerung zum…
#>  2 Neueste Daten OGD_bevjahresanf_PR_BEVJA_7          Bevölkerung zum…
#>  3 Neueste Daten OGD_bevjahresanf_PR_BEVJA_8          Bevölkerung zum…
#>  4 Neueste Daten OGD_bevjahresanfgebland_PR_BEVJAGB_4 Bevölkerung zum…
#>  5 Neueste Daten OGD_bevjahresanfgebland_PR_BEVJAGB_5 Bevölkerung zum…
#>  6 Neueste Daten OGD_bevjahresanfgebland_PR_BEVJAGB_7 Bevölkerung zum…
#>  7 Neueste Daten OGD_bevstprogjdgebland_PR_BEVJDGB_4  Bevölkerung zum…
#>  8 Neueste Daten OGD_bevstprogjdgebland_PR_BEVJDGB_5  Bevölkerung zum…
#>  9 Neueste Daten OGD_bevstprogjdgebland_PR_BEVJDGB_7  Bevölkerung zum…
#> 10 Neueste Daten OGD_bevstprog_PR_BEVJD_4             Bevölkerung zum…
#> # ℹ 374 more rows
subset(df, category == "Bildung und Forschung")
#> # A tibble: 45 × 3
#>    category              id                                    label  
#>    <chr>                 <chr>                                 <chr>  
#>  1 Bildung und Forschung OGD_persph_LEHR_1                     Lehrpe…
#>  2 Bildung und Forschung OGD_unistud5_ext_UNI_STUD5_1          Studie…
#>  3 Bildung und Forschung OGD_desi_hh_DESI_HH_1                 Indika…
#>  4 Bildung und Forschung OGD_innov015_CIS_015_Unt_Innovation_1 Untern…
#>  5 Bildung und Forschung OGD_innov014_CIS_014_Unt_Innovation_1 Umsätz…
#>  6 Bildung und Forschung OGD_innov013_CIS_013_Unt_Innovation_1 Innova…
#>  7 Bildung und Forschung OGD_innov013_CIS_013_Unt_Innovation_2 Innova…
#>  8 Bildung und Forschung OGD_innov012_CIS_012_Unt_Innovation_1 Untern…
#>  9 Bildung und Forschung OGD_innov011_CIS_011_Unt_Innovation_1 Untern…
#> 10 Bildung und Forschung OGD_fhsstud_ext_FHS_S_1               Studie…
#> # ℹ 35 more rows
# use an id to load a dataset
od_table("OGD_fhsstud_ext_FHS_S_1")
#> Studies at universities of applied sciences
#> 
#> Dataset: OGD_fhsstud_ext_FHS_S_1 (data.statistik.gv.at)
#> Measures: Ordinary Studies, Courses of studies (Lehrgang), Newly
#>   enrolled ordinary studies, Newly enrolled courses of studies
#>   (Lehrgang)
#> Fields: Semester <42>
#> 
#> Request: [2024-11-28 07:40:44.615657]
#> STATcubeR: 1.0.0