List available Opendata datasets

od_list() returns a data.frame containing all datasets published at data.statistik.gv.at

Usage

od_list(unique = TRUE, server = c("ext", "red"))

Arguments

unique

some datasets are pulbished under multiple groups. They will only be listed once with the first group they appear in unless this parameter is set to FALSE.

server

the open data server to use. Either ext for the external server (the default) or red for the editing server. The editing server is only accessible for employees of Statistics Austria

Value

a data.frame with two columns

"category": Grouping under which a dataset is listed

"id": Name of the dataset which can later be used in od_table()

"label": Description of the dataset

Examples

df <- od_list()
df
#> # A data frame: 315 × 3
#>    category    id                                        
#>    <chr>       <chr>                                     
#>  1 Arbeit      OGD_konjunkturmonitor_KonMon_1            
#>  2 Arbeit      OGD_veste309_Veste309_1                   
#>  3 Arbeit      OGD_veste307_Veste307_1                   
#>  4 Arbeit      OGD_veste305_Veste305_1                   
#>  5 Arbeit      OGD_veste303_Veste203_1                   
#>  6 Arbeit      OGD_veste301_Veste301_1                   
#>  7 Arbeit      OGD_aeapp_biber_abschl_ext_BIBER_ABSCHL_1 
#>  8 Arbeit      OGD_aest_erv_statw_haupt_ERV_STATWECHSEL_1
#>  9 Bevölkerung OGD_rate_kalwobez_GEST_KALWOCHE_STR_BZ_100
#> 10 Bevölkerung OGD_rate_kalwo_GEST_KALWOCHE_STR_100      
#> # … with 305 more rows, and 1 more variable: label <chr>
subset(df, category == "Bildung und Forschung")
#> # A data frame: 43 × 3
#>    category              id                                   
#>  * <chr>                 <chr>                                
#>  1 Bildung und Forschung OGD_innov015_CIS_015_Unt_Innovation_1
#>  2 Bildung und Forschung OGD_innov014_CIS_014_Unt_Innovation_1
#>  3 Bildung und Forschung OGD_innov012_CIS_012_Unt_Innovation_1
#>  4 Bildung und Forschung OGD_innov011_CIS_011_Unt_Innovation_1
#>  5 Bildung und Forschung OGD_innov013_CIS_013_Unt_Innovation_1
#>  6 Bildung und Forschung OGD_unistud1_ext_UNI_STUD1_1         
#>  7 Bildung und Forschung OGD_fhsstud_ext_FHS_S_1              
#>  8 Bildung und Forschung OGD_uptstud_ext_UPT_S_1              
#>  9 Bildung und Forschung OGD_uptabs_ext_UPT_A_1               
#> 10 Bildung und Forschung OGD_fhsabs_ext_FHS_A_1               
#> # … with 33 more rows, and 1 more variable: label <chr>
# use an id to load a dataset
od_table("OGD_fhsstud_ext_FHS_S_1")
#> An object of class od_table
#> 
#> Dataset     Studies at universities of applied sciences 
#> Measures    Ordinary Studies, Courses of studies (Lehrgang), Newly
#>             enrolled ordinary studies, Newly enrolled
#>             courses of studies (Lehrgang) 
#> Fields      Semester 
#> 
#> Request     2022-09-02 12:14:14 
#> STATcubeR   0.4.1