Resource management for open.data

Helper functions for caching and parsing open.data resources.

od_cache_dir(dir = NULL)

od_cache_clear(id)

od_cache_file(id, suffix = NULL, timestamp = NULL, ...)

od_resource(id, suffix = NULL, timestamp = NULL)

od_json(id, timestamp = Sys.time() - 3600)

od_resource_all(id, json = od_json(id))

Arguments

dir

If NULL, the cache directory is returned. Otherwise, the cache directory will be updated to dir.

id

A database id

suffix

A suffix for the resource: "HEADER" or a field code.

timestamp

A timestamp in POSIXct format. If provided, the cached resource will be updated if it is older than that value. Otherwise it will be downloaded only if it does not exist in the cache.

...

For internal use

json

The JSON file belonging to the dataset

Value

For od_cache_file() and od_resource(), the returned objects contain a hidden attribute attr(., "od") about the time used for downloading and parsing the resource. od_resource_all() converts these hidden attribute into columns.

Details

od_cache_clear(id) removes all files belonging to the specified id.

By default, downloaded json files will "expire" in one hour or 3600 seconds. That is, if a json is requested, it will be reused from the cache unless the file.mtime() is more than one hour behind Sys.time().

Examples

# get the current cache directory
od_cache_dir()
#> [1] "~/.cache/STATcubeR/open_data/"

# Get paths to cached files
od_cache_file("OGD_veste309_Veste309_1")
#> [1] "~/.cache/STATcubeR/open_data/OGD_veste309_Veste309_1.csv"
od_cache_file("OGD_veste309_Veste309_1", "C-A11-0")
#> [1] "~/.cache/STATcubeR/open_data/OGD_veste309_Veste309_1_C-A11-0.csv"

# get a parsed verison of the resource
od_resource("OGD_veste309_Veste309_1", "C-A11-0")
#> # A data frame: 3 × 4
#>   code  label label_de  label_en 
#> * <chr> <chr> <chr>     <chr>    
#> 1 A11-1 NA    insgesamt Sum total
#> 2 A11-2 NA    männlich  Male     
#> 3 A11-3 NA    weiblich  Female   

# Bundle all resources
od_resource_all("OGD_veste309_Veste309_1")
#> # A data frame: 6 × 7
#>   name      last_modified       cached               size download parsed data  
#>   <chr>     <dttm>              <dttm>              <dbl>    <dbl>  <dbl> <I<li>
#> 1 OGD_vest… 2021-03-24 11:32:41 2021-08-10 19:09:39  3857       NA  0.708 <tbl …
#> 2 OGD_vest… 2021-03-24 11:32:41 2021-08-10 19:09:39   516       NA  0.581 <tbl …
#> 3 OGD_vest… 2021-03-24 11:32:41 2021-08-10 19:09:39   159       NA  0.553 <tbl …
#> 4 OGD_vest… 2021-03-24 11:32:41 2021-08-10 19:09:39   697       NA  0.589 <tbl …
#> 5 OGD_vest… 2021-03-24 11:32:41 2021-08-10 19:09:39   518       NA  0.597 <tbl …
#> 6 OGD_vest… 2021-03-24 11:32:41 2021-08-10 19:09:39   641       NA  0.648 <tbl …