Get OGD revisions

Use the /revision endpoint of the OGD server to get a list of all datasets that have changed since a certain timestamp.

Usage

od_revisions(since = NULL, exclude_ext = TRUE, server = "ext")

Arguments

since

(optional) A timestamp. If supplied, only datasets updated later will be returned. Otherwise, all datasets are retured. Can be in either one of the following formats

a native R time type that is compatible with strftime() such as the return values of Sys.Date(), Sys.time() and file.mtime().

a string of the form YYYY-MM-DD to specify a day.

a string of the form YYYY-MM-DDThh:mm:ss to specify a day and a time.

exclude_ext

If TRUE (default) exclude all results that have OGDEXT_ as a prefix

server

the open data server to use. Either ext for the external server (the default) or red for the editing server. The editing server is only accessible for employees of Statistics Austria

Value

a character verctor with dataset ids

Examples

# get all datasets (including OGDEXT_*)
ids <- od_revisions(exclude_ext = FALSE)
length(ids)
#> [1] 370
head(ids)
#> [1] "OGDEXT_AEST_GEMTAB_1" "OGDEXT_AMB_1"        
#> [3] "OGDEXT_BINNENWAND_1"  "OGDEXT_DSR_1"        
#> [5] "OGDEXT_EHL_1"         "OGDEXT_FA_1"         

# get all the datasets since the last cache update
od_cache_dir() %>%
  file.mtime() %>%
  od_revisions()
#>  [1] "OGD_1905fue_FUE_B1905FUE_1"                      
#>  [2] "OGD_ake002q_AKEQ_1"                              
#>  [3] "OGD_beschaeftigtenindex2015_KJID2015_BI_1"       
#>  [4] "OGD_bkibrueckenbau2020_BKI_B20_1"                
#>  [5] "OGD_bkistrassebitumen2020_BKI_SB20_1"            
#>  [6] "OGD_bkistrassenbau2020_BKI_S20_1"                
#>  [7] "OGD_bkiwasserbau2020_BKI_W20_1"                  
#>  [8] "OGD_bkiwhsiedlg2020_BKI_WS20_1"                  
#>  [9] "OGD_bruttoverdiensteindex2015a_KJID2015_BVIa_1"  
#> [10] "OGD_einbuergerungen_EINBJ_1"                     
#> [11] "OGD_einbuergerungen_quart_EINB_1"                
#> [12] "OGD_f0760ext_OD_PkwGZL_1"                        
#> [13] "OGD_f1787_MZ_HWZ_Q_1"                            
#> [14] "OGD_f1898_F1898_3"                               
#> [15] "OGD_f1899_F1899_1"                               
#> [16] "OGD_f1900_F1900_1"                               
#> [17] "OGD_f1901_F1901_1"                               
#> [18] "OGD_f1903fue_FUE_B1903FUE_1"                     
#> [19] "OGD_f1904fue_FUE_B1904FUE_1"                     
#> [20] "OGD_f1906fue_FUE_B1906FUE_1"                     
#> [21] "OGD_fkfzul0759_OD_PkwNZL_1"                      
#> [22] "OGD_gest_kalwo_GEST_KALWOCHE_100"                
#> [23] "OGD_gest_kalwo_alter_GEST_KALWOCHE_5J_100"       
#> [24] "OGD_gestorbene_ext_ohmed_GES_1"                  
#> [25] "OGD_kjiarbeitsvolindex2015_KJID2015_AST_1"       
#> [26] "OGD_kjiauftragseingangsindex2015_KJID2015_AEGI_1"
#> [27] "OGD_kjiprodindex2015_KJID2015_PI_1"              
#> [28] "OGD_kjiumsatzindex2015_KJID2015_UI_1"            
#> [29] "OGD_klassen_KLASSEN_1"                           
#> [30] "OGD_konjunkturmonitor_KonMon_1"                  
#> [31] "OGD_krankenbewegungen_ex_ENTLASSUNGEN_AKUTKA_3"  
#> [32] "OGD_krankenbewegungen_ex_ENTLASSUNGEN_KA_2"      
#> [33] "OGD_krankenbewegungen_ex_LEISTUNGEN_1"           
#> [34] "OGD_lehrer_LEHR_1"                               
#> [35] "OGD_pregpi003_GHPI_20_1"                         
#> [36] "OGD_pregps003_GHPI_20S_1"                        
#> [37] "OGD_prodindexstd2015_KJID2015_PVISTD_1"          
#> [38] "OGD_prodvindexusb2015_KJID2015_PVISTD_1"         
#> [39] "OGD_rate_kalwo_GEST_KALWOCHE_STR_100"            
#> [40] "OGD_rate_kalwobez_GEST_KALWOCHE_STR_BZ_100"      
#> [41] "OGD_tli16kv_TLI_108"                             
#> [42] "OGD_tli16kvg_TLI_109"                            
#> [43] "OGD_tli16nace20_TLI_110"                         
#> [44] "OGD_tli16wk15_TLI_107"                           
#> [45] "OGD_vgr001_VGRJahresR_1"                         
#> [46] "OGD_vgr102_VGR_Besch_vj_1"                       
#> [47] "OGD_vgr103_VGR_HA_vj_1"                          
#> [48] "OGD_vgr104_VGR_Konsum_1"                         
#> [49] "OGD_vgr105_VGR_HA_Bws_1"                         
#> [50] "OGD_vgr106_VGR_HA_Prod_1"                        
#> [51] "OGD_vgr107_VGR_HA_BAI_1"                         
#> [52] "OGD_vgr108_VGR_HA_vj_1"                          
#> [53] "OGD_vgr110_VGR_Personen_1"                       
#> [54] "OGD_vpi00_VPI_2000_1"                            
#> [55] "OGD_vpi05_VPI_2005_1"                            
#> [56] "OGD_vpi10_VPI_2010_1"                            
#> [57] "OGD_vpi15_VPI_2015_1"                            
#> [58] "OGD_vpi20_VPI_2020_1"                            
#> [59] "OGD_vpi58_VPI_1958_i_1"                          
#> [60] "OGD_vpi58ii_VPI_1958_ii_1"                       
#> [61] "OGD_vpi66_VPI_1966_1"                            
#> [62] "OGD_vpi76_VPI_1976_1"                            
#> [63] "OGD_vpi86_VPI_1986_1"                            
#> [64] "OGD_vpi96_VPI_1996_1"                            
#> [65] "OGD_vpikhpi_VPI_KHPI_1"                          
#> [66] "OGD_vpilhki38_VPI_LHKI38_1"                      
#> [67] "OGD_vpilhki45_VPI_LHKI45_1"                      
#> [68] "OGD_zlf_komm_ZLF_KOM_1"                          
#> attr(,"reponse")
#> Response [https://data.statistik.gv.at/ogd/revision?since_time=2022-09-06T15%3A24%3A12]
#>   Date: 2022-09-19 16:30
#>   Status: 200
#>   Content-Type: application/json;charset=UTF-8
#>   Size: 2.56 kB
#> [
#>   "OGDEXT_AMB_1",
#>   "OGDEXT_DSR_1",
#>   "OGDEXT_FA_1",
#>   "OGDEXT_GBEZ_1",
#>   "OGDEXT_GEM_1",
#>   "OGDEXT_GEM_MP_1",
#>   "OGDEXT_LWPG_1",
#>   "OGDEXT_NUTS_1",
#>   "OGDEXT_ORT_1",
#> ...