Create a table-instance from an open-data dataset

od_table() returns an R6-class object containing all relevant data and metadata from data.statistik.gv.at.

$data_raw contains the contents of {id}.csv

$meta includes information from {id}_HEADER.csv

$field(i) contains information from {id}_{field_code}.csv

od_table(id, language = c("en", "de"))

Arguments

id
the name of the data-set that should be accessed

language
the language to be used for labelling data: "en" or "de"

Details

The returned abject also provides certain functionalities to label and aggreate the dataset.

object$data takes object$data_raw and applies labels based on the metadata.

object$tabulate() also applies labeling and aggregates the dataset. See od_tabulate() for more information.

object$language can be used to get or set the language of the dataset. This affects the labelling behavior

Examples

x <- od_table("OGD_krebs_ext_KREBS_1")

#> Error: HEADER and attribute description do not match

## metadata
x

#> Error in eval(expr, envir, enclos): object 'x' not found

x$meta

#> Error in eval(expr, envir, enclos): object 'x' not found

x$field("Sex")

#> Error in eval(expr, envir, enclos): object 'x' not found

x$field(3)

#> Error in eval(expr, envir, enclos): object 'x' not found

## data
x$data_raw %>% head()

#> Error in head(.): object 'x' not found

x$data     %>% head()

#> Error in head(.): object 'x' not found

## switch language
x$language <- "de"

#> Error in x$language <- "de": object 'x' not found

x

#> Error in eval(expr, envir, enclos): object 'x' not found

x$data     %>% head()

#> Error in head(.): object 'x' not found

## tabulation: see `?od_tabulate` for more examples
x$tabulate("Reporting year", "Sex") %>% head()

#> Error in head(.): object 'x' not found

## other interesting tables
od_table("OGD_veste309_Veste309_1")

#> Error: HEADER and attribute description do not match

od_table("OGD_konjunkturmonitor_KonMon_1")

#> Error: HEADER and attribute description do not match

od_table("OGD_krankenbewegungen_ex_LEISTUNGEN_1")

#> Error: HEADER and attribute description do not match

od_table("OGD_f1741_HH_Proj_1")

#> Error: HEADER and attribute description do not match

od_table("OGD_veste303_Veste203_1")

#> Error: HEADER and attribute description do not match