Other endpoints of the STATcube REST API

Utilize the simple endpoints /info and /table_rate_limit. Those provide information about available locales and the amount of requests available for calls against the /table endpoint.

sc_info(language = c("en", "de"), key = sc_key())

sc_rate_limit_table(language = c("en", "de"), key = sc_key())

Arguments

language

The language to be used for labeling. "en" or "de"

key

(string) An API key. To display your key, call sc_browse_preferences().

Value

sc_info() returns information about all available database languages

sc_rate_limit_table() returns a 3x1 dataframe with the following columns

remaining how much requests can be sent to the /table endpoint until the rate limit is reached.

limit the number of requests allowed per hour.

reset a tiestamp when the rate limit will be reset. Ususally, this should be less than one hour `after the current time.

Examples

if (sc_key_exists()) {

sc_info()
sc_rate_limit_table()

}
#> remaining: 100/100
#> reset:     2021-09-22 16:39:23