Common interface for STATcubeR datasets

This class represents a common interface for datasets returned from the STATcube REST API and OGD datasets. sc_data obects are usually created with od_table(), sc_table() or df_table().

Active bindings

data

the raw data. A data.frame that uses codes for all field variables and for all column names. To obtain labeled data, use $tabulate().

meta

A list containing metadata about the dataset. It has at least the following entries

$source is a dataframe with a single row that contains information on the data source.

$measures is a dataframe with one row for every measure in the dataset. It contains codes and labels for each measure as well as the number of NAs. Derived classes might add additional columns.

$fields is a dataframe with one row for every field in the dataset. It contains codes and labels for each measure as well as the total codes. Derived classes might add additional columns

Methods

Public methods

sc_data$new()

sc_data$field()

sc_data$tabulate()

sc_data$total_codes()

sc_data$clone()

Method new()

This class is not exported. Use od_table(), sc_table() or df_table() to initialize objects of class sc_data.

Usage

sc_data$new(data, meta, fields)

Arguments

data, meta, fields

raw data, metadata and field information. Do not use directly but initialize objects with sc_table, od_table() or df_table()

Method field()

get information about a specific field. The format of the reurn value is similar to $meta. A data.frame that includes codes and labels for each level of the field.

Usage

sc_data$field(i = 1)

Arguments

i

specifier for the field. Integer or character. If an interger is provided, it should match the row number in $meta$fields. If a character is provided, the field is matched using pmatch() on all available codes and labels.

Examples

x <- od_table("OGD_krebs_ext_KREBS_1")
x$field(1)
x$field("Sex")

Method tabulate()

create a tidy dataset. See sc_tabulate() for details.

Usage

sc_data$tabulate(...)

Arguments

...

arguments that are passed down to sc_tabulate()

Examples

x <- od_table("OGD_krebs_ext_KREBS_1")
x$tabulate("Reporting year", "Sex")

Method total_codes()

Usage

sc_data$total_codes(...)

Arguments

...

key value pairs to define the total codes. Key should be a field code and value a code from $field(i). If empty, it will return a data.frame with all specified total codes. Keys and values can also use labels instead of codes. See examples.

Examples

earnings <- od_table("OGD_veste309_Veste309_1")
earnings$total_codes(Sex = "Sum total", Citizenship = "Total",
                     Region = "Total", `Form of employment` = "Total")
earnings$tabulate("Form of employment")
earnings$tabulate("Sex", "Form of employment")

Method clone()

The objects of this class are cloneable with this method.

Usage

sc_data$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

## create a new sc_data object via od_table()
x <- od_table("OGD_krebs_ext_KREBS_1")

## show data
x$data

#> # A STATcubeR tibble: 45,059 x 5
#>    `C-TUM_ICD10_3ST-0` `C-BERJ-0` `C-BUNDESLAND-0` `C-KRE_GESCHLECHT-0` `F-KRE`
#>  * <fct>               <fct>      <fct>            <fct>                  <int>
#>  1 TUM_ICD10_3ST-C00   BERJ-1983  BUNDESLAND-1     GESCHLECHT-1               2
#>  2 TUM_ICD10_3ST-C00   BERJ-1983  BUNDESLAND-2     GESCHLECHT-1               8
#>  3 TUM_ICD10_3ST-C00   BERJ-1983  BUNDESLAND-2     GESCHLECHT-2               2
#>  4 TUM_ICD10_3ST-C00   BERJ-1983  BUNDESLAND-3     GESCHLECHT-1               6
#>  5 TUM_ICD10_3ST-C00   BERJ-1983  BUNDESLAND-3     GESCHLECHT-2               2
#>  6 TUM_ICD10_3ST-C00   BERJ-1983  BUNDESLAND-4     GESCHLECHT-1              12
#>  7 TUM_ICD10_3ST-C00   BERJ-1983  BUNDESLAND-4     GESCHLECHT-2               2
#>  8 TUM_ICD10_3ST-C00   BERJ-1983  BUNDESLAND-5     GESCHLECHT-1               4
#>  9 TUM_ICD10_3ST-C00   BERJ-1983  BUNDESLAND-6     GESCHLECHT-1               5
#> 10 TUM_ICD10_3ST-C00   BERJ-1983  BUNDESLAND-6     GESCHLECHT-2               2
#> # … with 45,049 more rows

## show metadata
x$meta

#> $source
#> # STATcubeR metadata: 1 x 6
#>   code                  label                                                   
#>   <chr>                 <chr>                                                   
#> 1 OGD_krebs_ext_KREBS_1 Cancer statistics by reporting year, province of reside…
#> # … with 4 more columns: 'label_de', 'label_en', 'requested', 'scr_version'
#> 
#> $measures
#> # STATcubeR metadata: 1 x 5
#>   code  label                     NAs
#>   <chr> <chr>                   <int>
#> 1 F-KRE Number of records F-KRE     0
#> # … with 2 more columns: 'label_de', 'label_en'
#> 
#> $fields
#> # STATcubeR metadata: 4 x 7
#>   code               label                   total_code
#>   <chr>              <chr>                   <chr>     
#> 1 C-TUM_ICD10_3ST-0  Tumore ICD/10 3-Steller NA        
#> 2 C-BERJ-0           Reporting year          NA        
#> 3 C-BUNDESLAND-0     Province of residence   NA        
#> 4 C-KRE_GESCHLECHT-0 Sex                     NA        
#> # … with 4 more columns: 'label_de', 'label_en', 'nitems', 'type'
#> 

## ------------------------------------------------
## Method `sc_data$field`
## ------------------------------------------------

x <- od_table("OGD_krebs_ext_KREBS_1")
x$field(1)

#> # STATcubeR metadata: 94 x 5
#>    code              label                         parsed                       
#>    <chr>             <chr>                         <chr>                        
#>  1 TUM_ICD10_3ST-C00 <C00> Bösartige Neubildung d… <C00> Bösartige Neubildung d…
#>  2 TUM_ICD10_3ST-C01 <C01> Bösartige Neubildung d… <C01> Bösartige Neubildung d…
#>  3 TUM_ICD10_3ST-C02 <C02> Bösartige Neubildung s… <C02> Bösartige Neubildung s…
#>  4 TUM_ICD10_3ST-C03 <C03> Bösartige Neubildung d… <C03> Bösartige Neubildung d…
#>  5 TUM_ICD10_3ST-C04 <C04> Bösartige Neubildung d… <C04> Bösartige Neubildung d…
#>  6 TUM_ICD10_3ST-C05 <C05> Bösartige Neubildung d… <C05> Bösartige Neubildung d…
#>  7 TUM_ICD10_3ST-C06 <C06> Bösartige Neubildung s… <C06> Bösartige Neubildung s…
#>  8 TUM_ICD10_3ST-C07 <C07> Bösartige Neubildung d… <C07> Bösartige Neubildung d…
#>  9 TUM_ICD10_3ST-C08 <C08> Bösartige Neubildung s… <C08> Bösartige Neubildung s…
#> 10 TUM_ICD10_3ST-C09 <C09> Bösartige Neubildung d… <C09> Bösartige Neubildung d…
#> # … with 84 more rows
#> # … with 2 more columns: 'label_de', 'label_en'

x$field("Sex")

#> # STATcubeR metadata: 2 x 5
#>   code         label  parsed
#>   <chr>        <chr>  <chr> 
#> 1 GESCHLECHT-1 male   male  
#> 2 GESCHLECHT-2 female female
#> # … with 2 more columns: 'label_de', 'label_en'

## ------------------------------------------------
## Method `sc_data$tabulate`
## ------------------------------------------------

x <- od_table("OGD_krebs_ext_KREBS_1")
x$tabulate("Reporting year", "Sex")

#> # A STATcubeR tibble: 72 x 3
#>    `Reporting year` Sex    `Number of records F-KRE`
#>  * <date>           <fct>                      <int>
#>  1 1983-01-01       male                       14492
#>  2 1983-01-01       female                     17478
#>  3 1984-01-01       male                       14798
#>  4 1984-01-01       female                     17456
#>  5 1985-01-01       male                       14555
#>  6 1985-01-01       female                     17448
#>  7 1986-01-01       male                       14626
#>  8 1986-01-01       female                     17241
#>  9 1987-01-01       male                       14862
#> 10 1987-01-01       female                     17845
#> # … with 62 more rows

## ------------------------------------------------
## Method `sc_data$total_codes`
## ------------------------------------------------

earnings <- od_table("OGD_veste309_Veste309_1")
earnings$total_codes(Sex = "Sum total", Citizenship = "Total",
                     Region = "Total", `Form of employment` = "Total")
earnings$tabulate("Form of employment")

#> # A STATcubeR tibble: 6 x 6
#>   `Form of employment`        `Arithmetic mea… `1st quartile` `2nd quartile (me…
#> * <fct>                                  <int>          <int>              <int>
#> 1 "Standard employment "                    19             13                 16
#> 2 "Non-standard employment (…               15             10                 13
#> 3 "Non-standard employment: …               16             11                 14
#> 4 "Non-standard employment: …               17             10                 13
#> 5 "Non-standard employment: …               12              9                 10
#> 6 "Non-standard employment: …               14             12                 14
#> # … with 2 more variables: 3rd quartile <int>, Number of employees <int>

earnings$tabulate("Sex", "Form of employment")

#> # A STATcubeR tibble: 12 x 7
#>    Sex    `Form of employment`  `Arithmetic mea… `1st quartile` `2nd quartile (…
#>  * <fct>  <fct>                            <int>          <int>            <int>
#>  1 Male   "Standard employment…               20             14               17
#>  2 Male   "Non-standard employ…               17             11               14
#>  3 Male   "Non-standard employ…               19             11               15
#>  4 Male   "Non-standard employ…               18             11               14
#>  5 Male   "Non-standard employ…               13              9               10
#>  6 Male   "Non-standard employ…               14             12               14
#>  7 Female "Standard employment…               16             11               14
#>  8 Female "Non-standard employ…               15             10               13
#>  9 Female "Non-standard employ…               15             11               14
#> 10 Female "Non-standard employ…               15             10               13
#> 11 Female "Non-standard employ…               12              9               10
#> 12 Female "Non-standard employ…               13             10               12
#> # … with 2 more variables: 3rd quartile <int>, Number of employees <int>