Create a request against the /schema endpoint

Invoke the /schema endpoint of the STATcube REST API. In case of sc_schema_catalogue(), recurse into all datasets and tables and return a nested list with ids and labels for all resources. For sc_schema_db(), recurse into all valuesets and return a list of all resources available tor the specific dataset. The return values can be displayed as a tree object.

sc_schema(
  resource_id = NULL,
  depth = NULL,
  language = c("en", "de"),
  key = sc_key()
)

# S3 method for sc_schema
print(x, limit = 30, value = FALSE, ...)

sc_schema_catalogue(depth = "folder", key = sc_key())

sc_schema_db(db_id, depth = "valueset", key = sc_key())

Arguments

resource_id
A resource identifier un uid format

depth
If provided, the request will recurse into the given level. For datasets, available options are NULL (no recursion), "folder", "field" and "valueset". For the catalogue, only NULL and "folder" are applicable.

language
The language to be used for labeling. "en" or "de"

key
(string) An API key. To display your key, call sc_browse_preferences().

x
object to be printed

limit
maximum number of entries to be printed

value
show resources of type VALUE?

...
ignored

db_id
a database id

Examples

my_catalogue <- sc_schema_catalogue()

## print
my_catalogue

#>                                                                  levelName
#> 1  Extern V9                                                              
#> 2   ¦--Examples                                                           
#> 3   ¦   ¦--Communes (Demo)                                                
#> 4   ¦   ¦--People                                                         
#> 5   ¦   ¦--People multilingual                                            
#> 6   ¦   °--Retail Banking ML en                                           
#> 7   ¦--Publications and services                                          
#> 8   ¦   ¦--Economic Trend Monitor                                         
#> 9   ¦   ¦   °--Economic Trend Monitor                                     
#> 10  ¦   °--Austrian Economic Atlas                                        
#> 11  ¦       ¦--01 Key data Austria and European Union                     
#> 12  ¦       ¦--02 Key data Federal provinces                              
#> 13  ¦       ¦--03 Key data on economic branches                           
#> 14  ¦       ¦--04 Key data on economic branches by enterprise size classes
#> 15  ¦       ¦--05 Key data on economic branches compared with the EU      
#> 16  ¦       ¦--06 Key data on economic branches by federal provinces      
#> 17  ¦       ¦--07 Key data industry                                       
#> 18  ¦       ¦--08 Key data construction                                   
#> 19  ¦       ¦--09 Key data trade                                          
#> 20  ¦       ¦--10 Key data tourism                                        
#> 21  ¦       ¦--11 Key data transport                                      
#> 22  ¦       ¦--12 Key data foreign trade                                  
#> 23  ¦       °--13 Key data agriculture and foresty                        
#> 24  °--Statistics                                                         
#> 25      ¦--Labour Market                                                  
#> 26      ¦   ¦--Job Vacancies                                              
#> 27      ¦   ¦   ¦--Offene Stellen (Jahresdurchschnitt) ab 2009            
#> 28      ¦   ¦   °--Offene Stellen Quartale ab 2009                        
#> 29      ¦   ¦--Working hours (Labour Force Survey)                        
#> 30      ¦   °--... 3 nodes w/ 0 sub                                       
#> 31      °--... 18 nodes w/ 608 sub                                        
#>        type
#> 1    FOLDER
#> 2    FOLDER
#> 3  DATABASE
#> 4  DATABASE
#> 5  DATABASE
#> 6  DATABASE
#> 7    FOLDER
#> 8    FOLDER
#> 9  DATABASE
#> 10   FOLDER
#> 11 DATABASE
#> 12 DATABASE
#> 13 DATABASE
#> 14 DATABASE
#> 15 DATABASE
#> 16 DATABASE
#> 17 DATABASE
#> 18 DATABASE
#> 19 DATABASE
#> 20 DATABASE
#> 21 DATABASE
#> 22 DATABASE
#> 23 DATABASE
#> 24   FOLDER
#> 25   FOLDER
#> 26   FOLDER
#> 27 DATABASE
#> 28 DATABASE
#> 29 DATABASE
#> 30         
#> 31         

## access the parsed catalogue
my_catalogue$Statistics$`Labour Market`

#>                                                        levelName     type
#> 1 Labour Market                                                    FOLDER
#> 2  ¦--Job Vacancies                                                FOLDER
#> 3  ¦   ¦--Offene Stellen (Jahresdurchschnitt) ab 2009            DATABASE
#> 4  ¦   °--Offene Stellen Quartale ab 2009                        DATABASE
#> 5  ¦--Working hours (Labour Force Survey)                        DATABASE
#> 6  ¦--Austrian Micro census - Labour Force Survey Yearly Data    DATABASE
#> 7  ¦--Austrian Micro census - Labour Force Survey Quarterly Data DATABASE
#> 8  °--Microcensus Time Series since 1974                         DATABASE

my_catalogue$Statistics$`Labour Market`$`Working hours (Labour Force Survey)`

#> $id
#> [1] "str:database:deake005"
#> 
#> $label
#> [1] "Working hours (Labour Force Survey)"
#> 
#> $location
#> [1] "http://sdbext:8082/statistik.at/ext/statcube/rest/v1/schema/str:database:deake005"
#> 
#> $type
#> [1] "DATABASE"
#> 
#> $lastModified
#> [1] 1.61338e+12
#> 
#> $recordViewEnabled
#> [1] FALSE
#> 

db_schema <- sc_schema_db("deake005")

# printing
db_schema

#>                                                                                   levelName
#> 1  Working hours (Labour Force Survey)                                                     
#> 2   ¦--Factors                                                                             
#> 3   ¦   ¦--Average hours actually worked per week                                          
#> 4   ¦   ¦   °--Average hours actually worked per week                                      
#> 5   ¦   ¦--Average hours usually worked per week                                           
#> 6   ¦   ¦   °--Average hours usually worked per week                                       
#> 7   ¦   ¦--Volume of hours worked in the main job per year in million hours                
#> 8   ¦   ¦   °--Volume of hours worked in the main job per year in million hours            
#> 9   ¦   ¦--Volume of hours worked overtime (paid) per year in million hours                
#> 10  ¦   ¦   °--Volume of hours worked overtime (paid) per year in million hours            
#> 11  ¦   ¦--Volume of hours worked overtime (unpaid) per year in million hours              
#> 12  ¦   ¦   °--Volume of hours worked overtime (unpaid) per year in million hours          
#> 13  ¦   ¦--Volume of hours worked overtime (paid/unpaid) per year in million hours         
#> 14  ¦   ¦   °--Volume of hours worked overtime (paid/unpaid) per year in million hours     
#> 15  ¦   ¦--Volume of hours worked in the second job per year in million hours              
#> 16  ¦   ¦   °--Volume of hours worked in the second job per year in million hours          
#> 17  ¦   ¦--Average hours worked overtime per week related to the total of employees        
#> 18  ¦   ¦   °--Average hours worked overtime per week related to the total of employees    
#> 19  ¦   °--Average hours worked overtime per week related to employees working overtime    
#> 20  ¦       °--Average hours worked overtime per week related to employees working overtime
#> 21  ¦--Facts                                                                               
#> 22  ¦   ¦--GEWTATJAHR                                                                      
#> 23  ¦   ¦   °--GEWTATJAHR                                                                  
#> 24  ¦   ¦--GEWJAHR                                                                         
#> 25  ¦   ¦   °--GEWJAHR                                                                     
#> 26  ¦   ¦--TATSTUNDMALGEW                                                                  
#> 27  ¦   ¦   °--TATSTUNDMALGEW                                                              
#> 28  ¦   ¦--STUNDMALGEW                                                                     
#> 29  ¦   ¦   °--STUNDMALGEW                                                                 
#> 30  ¦   °--... 13 nodes w/ 13 sub                                                          
#> 31  °--... 7 nodes w/ 105 sub                                                              
#>        type
#> 1  DATABASE
#> 2     GROUP
#> 3   MEASURE
#> 4       SUM
#> 5   MEASURE
#> 6       SUM
#> 7   MEASURE
#> 8       SUM
#> 9   MEASURE
#> 10      SUM
#> 11  MEASURE
#> 12      SUM
#> 13  MEASURE
#> 14      SUM
#> 15  MEASURE
#> 16      SUM
#> 17  MEASURE
#> 18      SUM
#> 19  MEASURE
#> 20      SUM
#> 21    GROUP
#> 22  MEASURE
#> 23      SUM
#> 24  MEASURE
#> 25      SUM
#> 26  MEASURE
#> 27      SUM
#> 28  MEASURE
#> 29      SUM
#> 30         
#> 31         

# access child nodes
db_schema$`Demographic Characteristics`

#>                                                                                          levelName
#> 1  Demographic Characteristics                                                                    
#> 2   ¦--Gender                                                                                     
#> 3   ¦   °--Gender                                                                                 
#> 4   ¦--Age in years-groups                                                                        
#> 5   ¦   ¦--Age in years-groups                                                                    
#> 6   ¦   ¦--Alter in Jahresgruppen                                                                 
#> 7   ¦   °--Alter (Ebene +1)                                                                       
#> 8   ¦--Highest level of training completed (national classification)                              
#> 9   ¦   ¦--Highest level of training completed (national classification)                          
#> 10  ¦   °--Höchste abgeschlossene Schulbildung - nationale Gliederung (Ebene +1)                  
#> 11  ¦--Highest level of training completed (ISCED 97)                                             
#> 12  ¦   ¦--Highest level of training completed (ISCED 97)                                         
#> 13  ¦   °--Höchste abgeschlossene Ausbildung (ISCED-97) (Ebene +1)                                
#> 14  ¦--Highest level of training completed (ISCED 2011)                                           
#> 15  ¦   ¦--Highest level of training completed (ISCED 2011)                                       
#> 16  ¦   °--Höchste abgeschlossene Schulbildung - internationale Gliederung (ISCED 2011) (Ebene +1)
#> 17  ¦--Nationality                                                                                
#> 18  ¦   ¦--Nationality                                                                            
#> 19  ¦   °--Staatsangehörigkeit                                                                    
#> 20  ¦--Country of birth                                                                           
#> 21  ¦   ¦--Country of birth                                                                       
#> 22  ¦   °--Geburtsland Österreich                                                                 
#> 23  °--Migration background                                                                       
#> 24      ¦--Migration background                                                                   
#> 25      °--Migrationshintergrund (Ebene +1)                                                       
#>        type
#> 1     GROUP
#> 2     FIELD
#> 3  VALUESET
#> 4     FIELD
#> 5  VALUESET
#> 6  VALUESET
#> 7  VALUESET
#> 8     FIELD
#> 9  VALUESET
#> 10 VALUESET
#> 11    FIELD
#> 12 VALUESET
#> 13 VALUESET
#> 14    FIELD
#> 15 VALUESET
#> 16 VALUESET
#> 17    FIELD
#> 18 VALUESET
#> 19 VALUESET
#> 20    FIELD
#> 21 VALUESET
#> 22 VALUESET
#> 23    FIELD
#> 24 VALUESET
#> 25 VALUESET

db_schema$`Demographic Characteristics`$Gender$Gender

#>                  levelName     type
#> 1 Gender                   VALUESET
#> 2  ¦--male                    VALUE
#> 3  ¦--female                  VALUE
#> 4  °--Not classifiable <0>    VALUE

db_schema$`Demographic Characteristics`$Gender$Gender$male

#> $id
#> [1] "str:value:deake005:F-DATA:C-C11-0:C-C11-0:C11-1"
#> 
#> $label
#> [1] "male"
#> 
#> $location
#> [1] "http://sdbext:8082/statistik.at/ext/statcube/rest/v1/schema/str:value:deake005:F-DATA:C-C11-0:C-C11-0:C11-1"
#> 
#> $type
#> [1] "VALUE"
#> 

# access the raw response from httr::GET()
my_response <- attr(db_schema, "response")
my_response$headers$date

#> [1] "Mon, 15 Feb 2021 09:17:52 GMT"

my_content <- httr::content(my_response)
my_content$label

#> [1] "Working hours (Labour Force Survey)"