Create a request against the /schema endpoint

Invoke the /schema endpoint of the STATcube REST API. In case of sc_schema_catalogue(), recurse into all datasets and tables and return a nested list with ids and labels for all resources. For sc_schema_db(), recurse into all valuesets and return a list of all resources available tor the specific dataset. The return values can be displayed as a tree object.

sc_schema(
  resource_id = NULL,
  depth = NULL,
  language = c("en", "de"),
  key = sc_key()
)

# S3 method for sc_schema
print(x, limit = 30, value = FALSE, recursive = TRUE, ...)

sc_schema_catalogue(depth = "folder", language = c("en", "de"), key = sc_key())

sc_schema_db(
  db_id,
  depth = "valueset",
  language = c("en", "de"),
  key = sc_key()
)

Arguments

resource_id
A resource identifier in uid format

depth
If provided, the request will recurse into the given level. For datasets, available options are NULL (no recursion), "folder", "field" and "valueset". For the catalogue, only NULL and "folder" are applicable.

language
The language to be used for labeling. "en" or "de"

key
(string) An API key. To display your key, call sc_browse_preferences().

x
object to be printed

limit
maximum number of entries to be printed

value
show resources of type VALUE?

recursive
should the whole tree be shown? Alternatively, only display the direct children of the current resource.

...
ignored

db_id
a database id

Examples

my_catalogue <- sc_schema_catalogue()

## print
my_catalogue

## access the parsed catalogue
my_catalogue$Statistics$`Labour Market`
my_catalogue$Statistics$`Labour Market`$`Working hours (Labour Force Survey)`

db_schema <- sc_schema_db("deake005")

# printing
db_schema

# access child nodes
db_schema$`Demographic Characteristics`
db_schema$`Demographic Characteristics`$Gender$Gender
db_schema$`Demographic Characteristics`$Gender$Gender$male

# access the raw response from httr::GET()
my_response <- attr(db_schema, "response")
my_response$headers$date
my_content <- httr::content(my_response)
my_content$label