Create a request against the /table endpoint

Send requests against the /table endpoint of the STATcube REST API. The requests can use three formats with corresponding functions

sc_table() uses a json file downloaded via the STATcube GUI

sc_table_custom() uses the ids of a database, measures and fields

sc_table_saved() uses a table uri of a saved table.

Those three functions all return an object of class "sc_table".

sc_table(
  json_file,
  language = c("en", "de", "both"),
  add_totals = TRUE,
  key = sc_key()
)

sc_examples_list()

sc_example(filename)

sc_table_custom(
  db,
  measures,
  dimensions,
  language = c("en", "de"),
  add_totals = TRUE,
  key = sc_key()
)

sc_table_saved_list()

sc_table_saved(table_uri, language = c("en", "de"), key = sc_key())

Arguments

json_file

path to a json file, which was downloaded via the STATcube GUI ("Open Data API Abfrage")

language

The language to be used for labeling. "en" or "de". The third option "both" will import both languages by sending two requests to the /table endpoint.

add_totals

Should totals be added for each measure in the json request?

key

(string) An API key. To display your key, call sc_browse_preferences().

filename

The name of an example json file.

db

The uid of a database

measures

A character vector of uids for measures. Can be either of type MEASURE or of type STAT_FUNCTION

dimensions

A character vector of dimensions for the cube. Can be either of type FIELD or type VALUESET. Those entries are referred to as fields in the parsed API response

table_uri

Identifier of a saved table as returned by sc_table_saved_list()

Value

An object of class sc_table which contains the return value of the httr::POST() request in obj$response. The object also provides member functions to parse this response object. See sc_table_class for the class documentation.

Examples

if (sc_key_exists()) {

my_table <- sc_table(json_file = sc_example("population_timeseries.json"))

# print
my_table

# get matadata for the table
my_table$meta

# get a data.frame
as.data.frame(my_table)

# get metadata for field 2
my_table$field(2)

sc_table_custom(
  db = "str:database:detouextregsai",
  measures = c(
    "str:statfn:detouextregsai:F-DATA1:F-ANK:SUM",
    "str:measure:detouextregsai:F-DATA1:F-UEB"
  ),
  dimensions = c(
    "str:field:detouextregsai:F-DATA1:C-SDB_TIT-0",
    "str:valueset:detouextregsai:F-DATA1:C-C93-2:C-C93SUM-0"
  )
)

# get the ids and labels of all saved tables
(saved_tables <- sc_table_saved_list())
table_uri <- saved_tables$id[1]

# get a table based on one of these ids
my_response <- sc_table_saved(table_uri)
as.data.frame(my_response)

}
#> # A STATcubeR tibble: 162 x 4
#>    `Tourism commune [pa… `Season/Tourism Mo… `Country of origin`  `Nights spent`
#>    <fct>                 <date>              <fct>                         <dbl>
#>  1 Burgenland <AT11>     2019-01-01          Vienna <01>                  249092
#>  2 Burgenland <AT11>     2019-01-01          Burgenland (beg.05/…          72667
#>  3 Burgenland <AT11>     2019-01-01          Carinthia (beg.05/0…          48346
#>  4 Burgenland <AT11>     2019-01-01          Lower Austria (beg.…         237369
#>  5 Burgenland <AT11>     2019-01-01          Upper Austria (beg.…          75659
#>  6 Burgenland <AT11>     2019-01-01          Salzburg (beg.05/03…          23909
#>  7 Burgenland <AT11>     2019-01-01          Styria (beg.05/03) …         140707
#>  8 Burgenland <AT11>     2019-01-01          Tyrol (beg.05/03) <…          15501
#>  9 Burgenland <AT11>     2019-01-01          Vorarlberg (beg.05/…           5099
#> 10 Burgenland <AT11>     2019-06-01          Vienna <01>                  421993
#> # … with 152 more rows