Create a request against the /table endpoint

Send requests against the /table endpoint of the STATcube REST API. The requests can use three formats with corresponding functions

sc_table() uses a json file downloaded via the STATcube GUI

sc_table_custom() uses the ids of a database, measures and fields

sc_table_saved() uses a table uri of a saved table.

Those three functions all return an object of class "sc_table".

sc_table(
  json_file,
  language = c("en", "de"),
  add_totals = TRUE,
  key = sc_key()
)

sc_examples_list()

sc_example(filename)

sc_table_custom(
  db,
  measures,
  dimensions,
  language = c("en", "de"),
  key = sc_key()
)

sc_table_saved_list()

sc_table_saved(table_uri, language = c("en", "de"), key = sc_key())

Arguments

json_file
path to a json file, which was downloaded via the STATcube gui ("Open Data API Abfrage")

language
The language to be used for labeling. "en" or "de"

add_totals
Should totals be added for each measure in the json request?

key
(string) An API key. To display your key, call sc_browse_preferences().

filename
The name of an example json file.

db
The uid of a database

measures
A character vector of uids for measures. Can be either of type MEASURE or of type STAT_FUNCTION

dimensions
A character vector of dimensions for the cube. Can be either of type FIELD or type VALUESET. Those entries are referred to as fields in the parsed API response

table_uri
Identifier of a saved table as returned by sc_table_saved_list()

Value

An object of class sc_table which contains the return value of the httr::POST() request in obj$response. The object also provides member functions to parse this response object. See sc_table_class for the class documentation.

See also

Examples

my_table <- sc_table(json_file = sc_example("population_timeseries.json"))

# print
my_table

# get matadata for the table
my_table$meta

# get a data.frame
as.data.frame(my_table) %>% head()

# get metadata for field 2
my_table$field(2)

sc_table_custom(
  db = "str:database:detouextregsai",
  measures = c(
    "str:statfn:detouextregsai:F-DATA1:F-ANK:SUM",
    "str:measure:detouextregsai:F-DATA1:F-UEB"
  ),
  dimensions = c(
    "str:field:detouextregsai:F-DATA1:C-SDB_TIT-0",
    "str:valueset:detouextregsai:F-DATA1:C-C93-2:C-C93SUM-0"
  )
)

# get the ids and labels of all saved tables
(saved_tables <- sc_table_saved_list())
table_uri <- saved_tables$id[1]

# get a table based on one of these ids
my_response <- sc_table_saved(table_uri)
as.data.frame(my_response) %>% head()