Create a request against the /table endpoint

Send requests against the /table endpoint of the STATcube REST API. The requests can use three formats with corresponding functions

sc_table() uses a json file downloaded via the STATcube GUI

sc_table_custom() uses the ids of a database, measures and fields

sc_table_saved() uses a table uri of a saved table.

Those three functions all return an object of class "sc_table".

sc_table(
  json_file,
  language = c("en", "de"),
  add_totals = TRUE,
  key = sc_key()
)

sc_examples_list()

sc_example(filename)

sc_table_custom(
  db,
  measures,
  dimensions,
  language = c("en", "de"),
  add_totals = TRUE,
  key = sc_key()
)

sc_table_saved_list()

sc_table_saved(table_uri, language = c("en", "de"), key = sc_key())

Arguments

json_file
path to a json file, which was downloaded via the STATcube gui ("Open Data API Abfrage")

language
The language to be used for labeling. "en" or "de"

add_totals
Should totals be added for each measure in the json request?

key
(string) An API key. To display your key, call sc_browse_preferences().

filename
The name of an example json file.

db
The uid of a database

measures
A character vector of uids for measures. Can be either of type MEASURE or of type STAT_FUNCTION

dimensions
A character vector of dimensions for the cube. Can be either of type FIELD or type VALUESET. Those entries are referred to as fields in the parsed API response

table_uri
Identifier of a saved table as returned by sc_table_saved_list()

Value

An object of class sc_table which contains the return value of the httr::POST() request in obj$response. The object also provides member functions to parse this response object. See sc_table_class for the class documentation.

See also

Examples

my_table <- sc_table(json_file = sc_example("population_timeseries.json"))

# print
my_table

#> An object of class sc_table
#> 
#> Database:      
#> Measures:      Number of persons 
#> Fields:        Quarter, Age in single years <96>, Sex <2>, Commune <2383> (Province-District) 
#> 
#> Request:       2021-08-11 10:53:31 
#> STATcubeR:     0.2.2

# get matadata for the table
my_table$meta

#> $source
#> # STATcubeR metadata: 1 x 3
#>   code       label                                                
#>   <chr>      <chr>                                                
#> 1 debevstand Population at the beginning of the quarter since 2002
#> # … with 1 more columns: 'scr_version'
#> 
#> $measures
#> # STATcubeR metadata: 1 x 6
#>   code     label               NAs
#>   <chr>    <chr>             <int>
#> 1 F-ISIS-1 Number of persons     0
#> # … with 3 more columns: 'fun', 'precision', 'annotations'
#> 
#> $fields
#> # STATcubeR metadata: 4 x 5
#>   code       label                              total_code
#>   <chr>      <chr>                              <chr>     
#> 1 C-A10-0    Quarter                            SC_TOTAL  
#> 2 C-BESC51-0 Age in single years <96>           SC_TOTAL  
#> 3 C-BESC11-0 Sex <2>                            SC_TOTAL  
#> 4 C-C41-2    Commune <2383> (Province-District) SC_TOTAL  
#> # … with 2 more columns: 'nitems', 'type'
#> 

# get a data.frame
as.data.frame(my_table) %>% head()

#> # A STATcubeR tibble: 6 x 5
#>   Quarter    `Age in single years <96>` `Sex <2>` `Commune <2383>… `Number of pers…
#> * <date>     <fct>                      <fct>     <fct>                       <dbl>
#> 1 2002-01-01 Up to 14 years old         male      Burgenland <AT1…            21287
#> 2 2002-01-01 Up to 14 years old         male      Carinthia <AT21>            47230
#> 3 2002-01-01 Up to 14 years old         male      Vienna <AT13>              117920
#> 4 2002-01-01 Up to 14 years old         male      Vorarlberg <AT3…            34798
#> 5 2002-01-01 Up to 14 years old         male      Tyrol <AT33>                62794
#> 6 2002-01-01 Up to 14 years old         male      Styria <AT22>               97538

# get metadata for field 2
my_table$field(2)

#> # STATcubeR metadata: 7 x 3
#>   code     label              parsed            
#>   <chr>    <chr>              <chr>             
#> 1 BESN07-1 Up to 14 years old Up to 14 years old
#> 2 BESN07-2 15 to 29 years old 15 to 29 years old
#> 3 BESN07-3 30 to 44 years old 30 to 44 years old
#> 4 BESN07-4 45 to 59 years old 45 to 59 years old
#> 5 BESN07-5 60 to 74 years old 60 to 74 years old
#> 6 BESN07-6 75 plus years old  75 plus years old 
#> 7 SC_TOTAL Total              Total             

sc_table_custom(
  db = "str:database:detouextregsai",
  measures = c(
    "str:statfn:detouextregsai:F-DATA1:F-ANK:SUM",
    "str:measure:detouextregsai:F-DATA1:F-UEB"
  ),
  dimensions = c(
    "str:field:detouextregsai:F-DATA1:C-SDB_TIT-0",
    "str:valueset:detouextregsai:F-DATA1:C-C93-2:C-C93SUM-0"
  )
)

#> An object of class sc_table
#> 
#> Database:      
#> Measures:      Arrivals, Nights spent 
#> Fields:        Season/Tourism Month, Country of origin 
#> 
#> Request:       2021-08-11 10:53:44 
#> STATcubeR:     0.2.2

# get the ids and labels of all saved tables
(saved_tables <- sc_table_saved_list())

#>               label                                             id
#> 1        ts_tourism str:table:cfc581ca-dd88-44e7-9ec2-cca153365dd5
#> 2 krankenbewegungen str:table:c7902e8d-5165-44e9-b17e-34ae20e2d1d4
#> 3        tourism_ts str:table:eec7dd70-25c4-4e5a-a6ae-1a9cd15d3c4c
#> 4      entlassungen str:table:f63f0713-155f-4d1d-8d41-4a50f0815fc7

table_uri <- saved_tables$id[1]

# get a table based on one of these ids
my_response <- sc_table_saved(table_uri)
as.data.frame(my_response) %>% head()

#> # A STATcubeR tibble: 6 x 4
#>   `Tourism commune [par… `Season/Tourism M… `Country of origin`   `Nights spent`
#> * <fct>                  <date>             <fct>                          <dbl>
#> 1 Burgenland <AT11>      2019-01-01         Vienna <01>                   249092
#> 2 Burgenland <AT11>      2019-01-01         Burgenland (beg.05/0…          72667
#> 3 Burgenland <AT11>      2019-01-01         Carinthia (beg.05/03…          48346
#> 4 Burgenland <AT11>      2019-01-01         Lower Austria (beg.0…         237369
#> 5 Burgenland <AT11>      2019-01-01         Upper Austria (beg.0…          75659
#> 6 Burgenland <AT11>      2019-01-01         Salzburg (beg.05/03)…          23909