Turn sc_data objects into tidy data frames

sc_tabulate() extracts the data in the table and turns it into a tidy data.frame. It applies labeling of the data and transforms time variables into a Date format if they satisfy certain STATcube Standards.

sc_tabulate(table, ...) is just an alias for table$tabulate(...) and was added so this rather complicated method can have a separate documentation page. It is recommended to use the table$tabulate() syntax

the ... argument decides which measures and/or fields should be included in the output. If no measures are given, all measures are included. The same is true for fields.

sc_tabulate(
  table,
  ...,
  .list = NULL,
  raw = FALSE,
  parse_time = TRUE,
  recode_zeros = inherits(table, "sc_table")
)

Arguments

table

An object of class sc_data

...

Names of measures and/or fields

.list

allows to define the arguments for ... as a character vector.

raw

If FALSE (the default), apply labeling to the dataset. Otherwise, return codes.

parse_time

Should time variables be converted into a Date format? Ignored if raw is set to TRUE.

recode_zeros

turn zero values into NAs

Details

Aggregation is done as follows

First, all columns that priovide a total code via table$total_codes() will be used to filter for column == total_code or column != total_code

Then, the remaining data is aggregated using rowsum()

The ellipsis (...) supports partial matching of codes and labels. See Examples

For objects of class sc_table two additional operations are performed.

zeros are recoded to NAs

rounding is done according to the precision of each measure. Ronding happens after the recoding to NA values

See also

Examples

############################ OGD Data #######################################

table <- od_table("OGD_veste309_Veste309_1")

# no arguments -> same output as `table$data`
table$tabulate()
#> # A STATcubeR tibble: 72 x 9
#>    Sex       Citizenship `Region (NUTS2)`   `Form of employmen… `Arithmetic mea…
#>  * <fct>     <fct>       <fct>              <fct>                          <int>
#>  1 Sum total Total       Total              "Total"                           18
#>  2 Sum total Total       Total              "Standard employme…               19
#>  3 Sum total Total       Total              "Non-standard empl…               15
#>  4 Sum total Total       Total              "Non-standard empl…               16
#>  5 Sum total Total       Total              "Non-standard empl…               17
#>  6 Sum total Total       Total              "Non-standard empl…               12
#>  7 Sum total Total       Total              "Non-standard empl…               14
#>  8 Sum total Total       AT11 Burgenland    "Total"                           16
#>  9 Sum total Total       AT12 Lower Austria "Total"                           16
#> 10 Sum total Total       AT13 Vienna        "Total"                           19
#> # … with 62 more rows, and 4 more variables: 1st quartile <int>,
#> #   2nd quartile (median) <int>, 3rd quartile <int>, Number of employees <int>

# provide some fields -> aggregate to keep only these fields
table$tabulate("Sex", "Citizenship")
#> # A STATcubeR tibble: 27 x 7
#>    Sex       Citizenship       `Arithmetic mea… `1st quartile` `2nd quartile (m…
#>  * <fct>     <fct>                        <int>          <int>             <int>
#>  1 Sum total "Total"                        267            182               231
#>  2 Sum total "Austrian"                      18             12                16
#>  3 Sum total "Non-Austrian "                 14             10                12
#>  4 Sum total "EU-28/EWR/Switz…               15             10                13
#>  5 Sum total "EU-15/EWR befor…               20             12                16
#>  6 Sum total "EU-13 EU Member…               13             10                12
#>  7 Sum total "Former Yugoslav…               13             10                12
#>  8 Sum total "Turkey"                        12              9                11
#>  9 Sum total "Other"                         14              9                11
#> 10 Male      "Total"                        290            198               245
#> # … with 17 more rows, and 2 more variables: 3rd quartile <int>,
#> #   Number of employees <int>

# provide some measures -> drop all other measures from the output
table$tabulate("Arithmetic mean")
#> # A STATcubeR tibble: 72 x 5
#>    Sex       Citizenship `Region (NUTS2)`   `Form of employmen… `Arithmetic mea…
#>  * <fct>     <fct>       <fct>              <fct>                          <int>
#>  1 Sum total Total       Total              "Total"                           18
#>  2 Sum total Total       Total              "Standard employme…               19
#>  3 Sum total Total       Total              "Non-standard empl…               15
#>  4 Sum total Total       Total              "Non-standard empl…               16
#>  5 Sum total Total       Total              "Non-standard empl…               17
#>  6 Sum total Total       Total              "Non-standard empl…               12
#>  7 Sum total Total       Total              "Non-standard empl…               14
#>  8 Sum total Total       AT11 Burgenland    "Total"                           16
#>  9 Sum total Total       AT12 Lower Austria "Total"                           16
#> 10 Sum total Total       AT13 Vienna        "Total"                           19
#> # … with 62 more rows

# mixture of measures and fields  -> keep exactly those columns
table$tabulate("Sex", "Arithmetic mean")
#> # A STATcubeR tibble: 3 x 2
#>   Sex       `Arithmetic mean`
#> * <fct>                 <int>
#> 1 Sum total               386
#> 2 Male                    417
#> 3 Female                  342

## define total codes
table$total_codes(
  `C-A11-0` = "A11-1",
  `C-STAATS-0` = "STAATS-9",
  `C-VEBDL-0` = "VEBDL-10",
  `C-BESCHV-0` = "BESCHV-1"
)

## alternatively, use partial matching to define totals
table$total_codes(
  Sex = "Sum total",
  Citizenship = "Total",
  Region = "Total",
  `Form of employment` = "Total"
)

# filter for totals in `Region (NUTS2)` and `Form of employment`. Drop totals
# in `Sex` and `Citizenship`.
table$tabulate("Sex", "Citizenship")
#> # A STATcubeR tibble: 16 x 7
#>    Sex    Citizenship          `Arithmetic mea… `1st quartile` `2nd quartile (m…
#>  * <fct>  <fct>                           <int>          <int>             <int>
#>  1 Male   "Austrian"                         20             14                17
#>  2 Male   "Non-Austrian "                    15             11                13
#>  3 Male   "EU-28/EWR/Switzerl…               16             11                14
#>  4 Male   "EU-15/EWR before 2…               21             13                17
#>  5 Male   "EU-13 EU Member St…               14             10                13
#>  6 Male   "Former Yugoslavia …               14             11                14
#>  7 Male   "Turkey"                           13             10                13
#>  8 Male   "Other"                            14              9                11
#>  9 Female "Austrian"                         16             11                14
#> 10 Female "Non-Austrian "                    13              9                11
#> 11 Female "EU-28/EWR/Switzerl…               14              9                11
#> 12 Female "EU-15/EWR before 2…               17             11                15
#> 13 Female "EU-13 EU Member St…               12              9                10
#> 14 Female "Former Yugoslavia …               11              9                10
#> 15 Female "Turkey"                           10              9                10
#> 16 Female "Other"                            13              9                11
#> # … with 2 more variables: 3rd quartile <int>, Number of employees <int>

## switch language
table$language <- "de"

## `...` matches for codes and labels
table$tabulate("C-A11-0", "Staats", "2. Quartil (Median)")
#> # A STATcubeR tibble: 16 x 3
#>    Geschlecht Staatsangehörigkeit                         `2. Quartil (Median)`
#>  * <fct>      <fct>                                                       <int>
#>  1 männlich   "Österreich"                                                   17
#>  2 männlich   "Nicht-Österreich"                                             13
#>  3 männlich   "EU-28/EWR/Schweiz (ohne Österreich)"                          14
#>  4 männlich   "EU-15/EWR vor 2004/Schweiz "                                  17
#>  5 männlich   "EU-13 Beitrittsstaaten ab 2004"                               13
#>  6 männlich   "ehemaliges Jugoslawien (außerhalb der EU)"                    14
#>  7 männlich   "Türkei"                                                       13
#>  8 männlich   "Andere"                                                       11
#>  9 weiblich   "Österreich"                                                   14
#> 10 weiblich   "Nicht-Österreich"                                             11
#> 11 weiblich   "EU-28/EWR/Schweiz (ohne Österreich)"                          11
#> 12 weiblich   "EU-15/EWR vor 2004/Schweiz "                                  15
#> 13 weiblich   "EU-13 Beitrittsstaaten ab 2004"                               10
#> 14 weiblich   "ehemaliges Jugoslawien (außerhalb der EU)"                    10
#> 15 weiblich   "Türkei"                                                       10
#> 16 weiblich   "Andere"                                                       11

## Keep totals in the output by removing total codes
table$tabulate("C-A11-0")      # -> 2 rows: "male" "female"
#> # A STATcubeR tibble: 2 x 6
#>   Geschlecht `Arithmetisches Mittel` `1. Quartil` `2. Quartil (Med… `3. Quartil`
#> * <fct>                        <int>        <int>             <int>        <int>
#> 1 männlich                        19           13                16           22
#> 2 weiblich                        15           10                13           18
#> # … with 1 more variable: Zahl d unselbst Beschäftigten <int>
table$total_codes(`C-A11-0` = NA)
table$tabulate("C-A11-0")      # -> 3 rows: "total", "male", "female"
#> # A STATcubeR tibble: 3 x 6
#>   Geschlecht `Arithmetisches Mittel` `1. Quartil` `2. Quartil (Med… `3. Quartil`
#> * <fct>                        <int>        <int>             <int>        <int>
#> 1 insgesamt                       18           12                15           20
#> 2 männlich                        19           13                16           22
#> 3 weiblich                        15           10                13           18
#> # … with 1 more variable: Zahl d unselbst Beschäftigten <int>

## table$tabulate(...) is an alias for sc_tabulate(table, ...)
sc_tabulate(table, "C-A11-0")
#> # A STATcubeR tibble: 3 x 6
#>   Geschlecht `Arithmetisches Mittel` `1. Quartil` `2. Quartil (Med… `3. Quartil`
#> * <fct>                        <int>        <int>             <int>        <int>
#> 1 insgesamt                       18           12                15           20
#> 2 männlich                        19           13                16           22
#> 3 weiblich                        15           10                13           18
#> # … with 1 more variable: Zahl d unselbst Beschäftigten <int>

######################### STATcube REST API #################################

if (sc_key_exists()) {

table_tourism <- sc_example("accomodation.json") %>% sc_table("de")

table_tourism$tabulate()
table_tourism$tabulate("Saison/Tourismusmonat")
table_tourism$tabulate("Saison/Tourismusmonat", "Ankünfte")
table_tourism$tabulate("Ankünfte")

## TODO: param annotations does not work currently
if (FALSE) {
  table_trade <- sc_example("foreign_trade.json") %>% sc_table("de")
  tt <- sc_tabulate(table_trade, "Berichtsjahr", "Import, Wert in Euro",
                    annotations = TRUE)
  tt
  tt[['Import, Wert in Euro_a']] %>% str()
}

}
#> Error in sc_key_exists() could not find function "sc_key_exists"