Turn sc_table objects into tidy data frames

sc_tabulate() extracts the data in the table and turns it into a tidy data.frame. Additionaly, certain STATcube conventions are applied

zeros are recoded to NAs

rounding is done according to the precision of each measure

time variables are converted into POSIXct

the ... argument decides which measures and/or fields should be included in the output. If no measures are given, all measures are included. The same is true for fields.

sc_tabulate(
  table,
  ...,
  .list = NULL,
  parse_time = TRUE,
  round = TRUE,
  recode_zeros = TRUE,
  annotations = FALSE,
  raw = FALSE
)

Arguments

table
An object of class sc_table

...
Names of measures and/or fields

.list
allows to define the arguments for ... as a character vector.

parse_time
should field variables of type time be converted into a POSIXct format?

round
use the precision of each measure for rounding?

recode_zeros
turn zero values into NAs prior to rounding

annotations
Include separate annotation columns in the returned data.frame? Those extra columns are of type list and contain character vectors with all annotations for the corresponding measure

raw
If FALSE (the default), apply labeling to the dataset. Otherwise, return codes.

Examples

table_tourism <- sc_example("accomodation.json") %>% sc_table("de")

sc_tabulate(table_tourism)
sc_tabulate(table_tourism, "Saison/Tourismusmonat")
sc_tabulate(table_tourism, "Saison/Tourismusmonat", "Ankünfte")
sc_tabulate(table_tourism, "Ankünfte")

table_trade <- sc_example("foreign_trade.json") %>% sc_table("de")
tt <- sc_tabulate(table_trade, "Berichtsjahr", "Import, Wert in Euro",
                  annotations = TRUE)
tt
tt[['Import, Wert in Euro_a']] %>% str()